refactor: split service.cpp into focused translation units
Addresses issue #66 review feedback. The monolithic service.cpp (≈730 loc)
is replaced by six focused files:

  service.cpp               — constructor + seedConfiguredKeyspaces bootstrap
  service_bootstrap.cpp     — ensureAllocators, seedConfiguredKeyspaces helpers
  service_work.cpp          — handleWork, handleHeartbeat, worker/chunk helpers
  service_submit.cpp        — handleSubmit, handleSubmitDone, handleSubmitFound
  service_admin.cpp         — handleActivatePuzzle, handleSetPuzzle, handleSetTestChunk, handleAdminPuzzles
  service_stats.cpp         — handleStats, buildStats
  service_serialization.cpp — puzzleJson, formatDouble, nowIsoUtc, jsonResponse helpers

CMakeLists.txt updated to list all seven translation units.
Clean rebuild verified: stale server.cpp.o artifact removed.
Also removes leftover server.js and test.sh from the old Node.js era.
