+++ deps.txt
@@ -1,5 +1,5 @@
 # Linux
-sudo apt install -y build-essential cmake ninja-build git pkg-config libsqlite3-dev libboost-all-dev nlohmann-json3-dev
+sudo apt install -y build-essential cmake ninja-build git pkg-config libsqlite3-dev libboost-all-dev nlohmann-json3-dev libasio-dev
 
 #mac
 brew update && brew install git cmake ninja pkg-config boost nlohmann-json sqlite asio
