--- server.js
@@ -1,2353 +1,0 @@
-const express = require('express');
-const Database = require('better-sqlite3');
-const crypto = require('crypto');
-const fs = require('fs');
-
-// --- Configuration ---
-const PORT            = parseInt(process.env.PORT || '8888', 10);
-const DB_PATH         = process.env.DB_PATH || 'pool.db';
-const TARGET_MINUTES  = parseInt(process.env.TARGET_MINUTES || '10', 10);
-const TIMEOUT_MINUTES = parseInt(process.env.TIMEOUT_MINUTES || '15', 10);
-const TARGET_SECTORS  = BigInt(parseInt(process.env.TARGET_SECTORS || '65536', 10));
-
-// Dashboard freshness threshold only. Green = recent heartbeat; gray = visible but stale.
-const DEFAULT_ACTIVE_MINUTES = 1.167;
-const rawActiveMinutes = parseFloat(process.env.ACTIVE_MINUTES || String(DEFAULT_ACTIVE_MINUTES));
-const ACTIVE_MINUTES_REQUESTED =
-    Number.isFinite(rawActiveMinutes) && rawActiveMinutes > 0 ? rawActiveMinutes : DEFAULT_ACTIVE_MINUTES;
-const ACTIVE_MINUTES = Math.max(0.1, Math.min(ACTIVE_MINUTES_REQUESTED, TIMEOUT_MINUTES / 2));
-
-const REACTIVATE_MINUTES = TIMEOUT_MINUTES;
-
-// --- Allocator strategies ---
-const ALLOC_STRATEGY_LEGACY  = 'legacy_random_shards_v1';
-const ALLOC_STRATEGY_VCHUNKS = 'virtual_random_chunks_v1';
-const DEFAULT_ALLOC_STRATEGY = process.env.DEFAULT_ALLOC_STRATEGY || ALLOC_STRATEGY_VCHUNKS;
-
-// 1 virtual chunk should be ~1 minute on the slowest client.
-// Default assumes ~500k keys/s slowest worker => 30,000,000 keys.
-const DEFAULT_VIRTUAL_CHUNK_SIZE_KEYS = BigInt(process.env.DEFAULT_VIRTUAL_CHUNK_SIZE_KEYS || '30000000');
-
-// Optional safety valve for existing DBs:
-// if enabled, puzzles using virtual chunks and having NO issued non-test chunks yet
-// are automatically re-seeded to the current default virtual chunk size on startup /
-// first use. This is useful when an older test DB still contains oversized chunks.
-const AUTO_RESEED_EMPTY_VCHUNK_PUZZLES = process.env.AUTO_RESEED_EMPTY_VCHUNK_PUZZLES === '1';
-
-// Virtual chunk universe is mathematical, not materialized in DB.
-// Still keep counts/indexes within JS safe integer range because they are
-// stored in SQLite INTEGER columns and read back into JS numbers.
-const MAX_STORED_VCHUNKS = BigInt(Number.MAX_SAFE_INTEGER);
-
-// Allocation does not scan the full universe. It probes a bounded number of
-// pseudo-random candidate positions each request.
-const MAX_ALLOC_PROBES = parseInt(process.env.MAX_ALLOC_PROBES || '8192', 10);
-
-const TARGET_MAX_SHARDS = BigInt(parseInt(process.env.TARGET_MAX_SHARDS || '65536', 10));
-const TARGET_MAX_VCHUNKS_PER_SHARD = parsePositiveBigInt(
-    process.env.TARGET_MAX_VCHUNKS_PER_SHARD,
-    null
-);
-
-// Puzzle #71 range floor for shard sizing:
-// start = 0x0400000000000000000
-// end   = 0x07fffffffffffffffff
-// range = end - start
-const PUZZLE71_RANGE_FLOOR = BigInt('1180591620717411303423');
-
-// GPU batch size kept for admin test-chunk default
-const GPU_BATCH_KEYS = 4278190080n;
-
-const PERMUTATION_MODE_RAW = process.env.PERMUTATION_MODE || 'feistel';
-const PERMUTATION_MODE =
-    PERMUTATION_MODE_RAW === 'affine' || PERMUTATION_MODE_RAW === 'feistel'
-        ? PERMUTATION_MODE_RAW
-        : (() => { throw new Error(`Invalid PERMUTATION_MODE: ${PERMUTATION_MODE_RAW}`); })();
-
-// --- Pure helpers ---
-
-function isValidHex(s) {
-    return typeof s === 'string' && /^(0x)?[0-9a-fA-F]{1,64}$/.test(s);
-}
-
-function normalizeHex(s) {
-    return BigInt('0x' + s.replace(/^0x/i, '')).toString(16).padStart(64, '0');
-}
-
-function randomBigIntInRange(min, max) {
-    const range = max - min;
-    if (range <= 0n) return min;
-    const hexLen = range.toString(16).length;
-    const byteLen = Math.ceil(hexLen / 2) + 8;
-    const bytes = crypto.randomBytes(byteLen);
-    const n = BigInt('0x' + bytes.toString('hex'));
-    return min + (n % range);
-}
-
-function encodeBigIndex(n) {
-    const x = typeof n === 'bigint' ? n : BigInt(n);
-    if (x < 0n) throw new Error(`encodeBigIndex: negative value ${x.toString()}`);
-    return x.toString(16).padStart(64, '0');
-}
-
-function decodeBigIndex(value, fallback = null) {
-    if (value === undefined || value === null || value === '') return fallback;
-
-    if (typeof value === 'bigint') return value;
-
-    if (typeof value === 'number') {
-        if (!Number.isFinite(value) || value < 0) return fallback;
-        return BigInt(Math.trunc(value));
-    }
-
-    const s = String(value).trim();
-    if (!s) return fallback;
-
-    if (/^0x[0-9a-fA-F]+$/.test(s)) return BigInt(s);
-    if (/^[0-9a-fA-F]{64}$/.test(s)) return BigInt('0x' + s);
-    if (/^[0-9]+$/.test(s)) return BigInt(s);
-
-    throw new Error(`decodeBigIndex: unsupported value ${value}`);
-}
-
-function normalizeRunStartForCandidateBig(candidateIndex, neededChunks, totalChunks) {
-    if (neededChunks >= totalChunks) return 0n;
-
-    let start = candidateIndex;
-    if (start + neededChunks > totalChunks) {
-        start = totalChunks - neededChunks;
-    }
-    if (start < 0n) start = 0n;
-    return start;
-}
-
-function sumShardLocalRuns(rows) {
-    let total = 0n;
-
-    for (const r of rows) {
-        const s = decodeBigIndex(r.local_vchunk_start, decodeBigIndex(r.vchunk_start, null));
-        const e = decodeBigIndex(r.local_vchunk_end, decodeBigIndex(r.vchunk_end, null));
-        if (s !== null && e !== null && e > s) {
-            total += (e - s);
-        }
-    }
-
-    return total.toString();
-}
-
-function normalizeHashrate(input, fallback = 1000000) {
-    const n = Number(input);
-    if (!Number.isFinite(n) || n <= 0) return BigInt(fallback);
-    return BigInt(Math.max(1, Math.floor(n)));
-}
-
-function sha256Hex(s) {
-    return crypto.createHash('sha256').update(String(s)).digest('hex');
-}
-
-function hmacSha256Hex(key, msg) {
-    return crypto.createHmac('sha256', String(key)).update(String(msg)).digest('hex');
-}
-
-function bigIntMin(a, b) {
-    return a < b ? a : b;
-}
-
-function bigIntMax(a, b) {
-    return a > b ? a : b;
-}
-
-function ceilDiv(a, b) {
-    return (a + b - 1n) / b;
-}
-
-function bitLengthBigInt(n) {
-    let x = n < 0n ? -n : n;
-    let bits = 0;
-    while (x > 0n) {
-        x >>= 1n;
-        bits++;
-    }
-    return bits;
-}
-
-function nextPowerOfTwoBigInt(n) {
-    if (n <= 1n) return 1n;
-    let p = 1n;
-    while (p < n) p <<= 1n;
-    return p;
-}
-
-function chooseShardSize(range) {
-    const rawShardSize = ceilDiv(range, TARGET_MAX_SHARDS);
-    const minShardSize = PUZZLE71_RANGE_FLOOR;
-    return nextPowerOfTwoBigInt(rawShardSize > minShardSize ? rawShardSize : minShardSize);
-}
-
-function chooseVirtualChunkSizeForShard(shardSize) {
-    let size = DEFAULT_VIRTUAL_CHUNK_SIZE_KEYS > 0n ? DEFAULT_VIRTUAL_CHUNK_SIZE_KEYS : 1n;
-    if (size > shardSize) size = shardSize;
-
-    if (TARGET_MAX_VCHUNKS_PER_SHARD && TARGET_MAX_VCHUNKS_PER_SHARD > 0n) {
-        while (ceilDiv(shardSize, size) > TARGET_MAX_VCHUNKS_PER_SHARD) {
-            size <<= 1n;
-            if (size > shardSize) {
-                size = shardSize;
-                break;
-            }
-        }
-    }
-
-    return size;
-}
-
-function validateShardCountOrThrow(shardCountBig, context) {
-    if (shardCountBig > BigInt(Number.MAX_SAFE_INTEGER)) {
-        throw new Error(
-            `${context}: shard count ${shardCountBig.toString()} exceeds JS safe integer range ` +
-            `(${Number.MAX_SAFE_INTEGER}).`
-        );
-    }
-}
-
-function shardLocalVirtualChunkRangeToHex(puzzle, shardIndex, localVChunkStart, localVChunkEndExclusive) {
-    const pStart = BigInt('0x' + puzzle.start_hex);
-    const pEnd   = BigInt('0x' + puzzle.end_hex);
-
-    const shardSize = BigInt(puzzle.shard_size_keys);
-    const vchunkSize = BigInt(puzzle.virtual_chunk_size_keys);
-
-    const absoluteStart = pStart
-        + BigInt(shardIndex) * shardSize
-        + BigInt(localVChunkStart) * vchunkSize;
-
-    const absoluteEnd = pStart
-        + BigInt(shardIndex) * shardSize
-        + BigInt(localVChunkEndExclusive) * vchunkSize;
-
-    const clippedEnd = bigIntMin(absoluteEnd, pEnd);
-
-    return {
-        startHex: absoluteStart.toString(16).padStart(64, '0'),
-        endHex: clippedEnd.toString(16).padStart(64, '0'),
-    };
-}
-
-function nextEven(n) {
-    return (n % 2 === 0) ? n : (n + 1);
-}
-
-function feistelRoundValue(right, roundKey, mask) {
-    const hex = hmacSha256Hex(roundKey, right.toString());
-    return BigInt('0x' + hex) & mask;
-}
-
-function permutePow2Feistel(x, bits, key) {
-    const halfBits = bits / 2;
-    const halfMask = (1n << BigInt(halfBits)) - 1n;
-
-    let left  = (x >> BigInt(halfBits)) & halfMask;
-    let right = x & halfMask;
-
-    const ROUNDS = 6;
-
-    for (let round = 0; round < ROUNDS; round++) {
-        const f = feistelRoundValue(
-            right,
-            `${key}:round:${round}`,
-            halfMask
-        );
-        const newLeft = right;
-        const newRight = (left ^ f) & halfMask;
-        left = newLeft;
-        right = newRight;
-    }
-
-    return ((left & halfMask) << BigInt(halfBits)) | (right & halfMask);
-}
-
-function permuteIndexFeistel(orderIndex, n, key) {
-    if (n <= 1n) return 0n;
-    if (orderIndex < 0n || orderIndex >= n) {
-        throw new Error(`permuteIndexFeistel: orderIndex ${orderIndex.toString()} out of range for n=${n.toString()}`);
-    }
-
-    const bits = nextEven(Math.max(1, bitLengthBigInt(n - 1n)));
-    let x = orderIndex;
-
-    // Cycle walking on the enclosing power-of-two domain.
-    for (;;) {
-        x = permutePow2Feistel(x, bits, key);
-        if (x < n) return x;
-    }
-}
-
-function roundUpToQuantum(value, quantum) {
-    if (quantum <= 1n) return value;
-    return ceilDiv(value, quantum) * quantum;
-}
-
-function parsePositiveBigInt(value, fallback = null) {
-    if (value === undefined || value === null || String(value) === '') return fallback;
-    try {
-        const n = BigInt(String(value));
-        if (n <= 0n) return fallback;
-        return n;
-    } catch (_) {
-        return fallback;
-    }
-}
-
-function normalizedRange(startHex, endHex) {
-    const start = BigInt('0x' + startHex);
-    const end   = BigInt('0x' + endHex);
-    if (end <= start) throw new Error('Puzzle range must be > 0');
-    return { start, end, range: end - start };
-}
-
-function gcdBigInt(a, b) {
-    a = a < 0n ? -a : a;
-    b = b < 0n ? -b : b;
-    while (b !== 0n) {
-        const t = a % b;
-        a = b;
-        b = t;
-    }
-    return a;
-}
-
-function deriveAffinePermutationParams(seedHex, n) {
-    if (n <= 1n) return { a: 1n, b: 0n };
-
-    let counter = 0n;
-    let a = 1n;
-    while (true) {
-        const ha = sha256Hex(`${seedHex}:a:${counter.toString()}`);
-        a = BigInt('0x' + ha) % n;
-        if (a === 0n) a = 1n;
-        if (gcdBigInt(a, n) === 1n) break;
-        counter++;
-    }
-
-    const hb = sha256Hex(`${seedHex}:b`);
-    const b = BigInt('0x' + hb) % n;
-
-    return { a, b };
-}
-
-function permuteIndexAffine(orderIndex, n, a, b) {
-    return (a * orderIndex + b) % n;
-}
-
-function defaultAllocSeedForPuzzle(puzzle, strategy = ALLOC_STRATEGY_VCHUNKS) {
-    return sha256Hex(`${puzzle.name}|${puzzle.start_hex}|${puzzle.end_hex}|${strategy}`);
-}
-
-function pickShardIndex(orderIndex, shardCountBig, permKey) {
-    if (shardCountBig <= 1n) return 0;
-    if (PERMUTATION_MODE === 'affine') {
-        const affine = deriveAffinePermutationParams(`${permKey}:shards`, shardCountBig);
-        return Number(permuteIndexAffine(orderIndex, shardCountBig, affine.a, affine.b));
-    }
-    return Number(permuteIndexFeistel(orderIndex, shardCountBig, `${permKey}:shards`));
-}
-
-function validateVirtualChunkCountOrThrow(chunkCountBig, context) {
-    if (chunkCountBig > MAX_STORED_VCHUNKS) {
-        throw new Error(
-            `${context}: virtual chunk count ${chunkCountBig.toString()} exceeds MAX_STORED_VCHUNKS ` +
-            `(${MAX_STORED_VCHUNKS.toString()}). Use a larger virtual_chunk_size_keys value.`
-        );
-    }
-}
-
-function computeWorkerRequestedKeys({ hashrateBig, minChunkKeys, chunkQuantumKeys }) {
-    const targetKeys = hashrateBig * BigInt(TARGET_MINUTES * 60);
-    const minKeys = minChunkKeys && minChunkKeys > 0n ? minChunkKeys : targetKeys;
-    const raw = bigIntMax(targetKeys, minKeys);
-    const quantum = chunkQuantumKeys && chunkQuantumKeys > 0n ? chunkQuantumKeys : 1n;
-    return roundUpToQuantum(raw, quantum);
-}
-
-function parseUtcDateMs(value) {
-    if (!value || typeof value !== 'string') return null;
-    let s = value.trim();
-    if (!s) return null;
-    if (!s.includes('T')) s = s.replace(' ', 'T');
-    if (!/[zZ]$|[+\-]\d{2}:\d{2}$/.test(s)) s += 'Z';
-    const ms = Date.parse(s);
-    return Number.isFinite(ms) ? ms : null;
-}
-
-function computeWorkerProgressPercent(assignedAt, hashrate, jobKeys) {
-    if (!assignedAt || !hashrate || !jobKeys) return null;
-    const assignedMs = parseUtcDateMs(assignedAt);
-    if (assignedMs === null) return null;
-    const elapsedSeconds = (Date.now() - assignedMs) / 1000;
-    if (elapsedSeconds < 0) return null;
-    const hashrateNum = Number(hashrate);
-    if (!Number.isFinite(hashrateNum) || hashrateNum <= 0) return null;
-    const jobKeysBig = typeof jobKeys === 'string' ? BigInt(jobKeys) : BigInt(Math.round(Number(jobKeys)));
-    if (jobKeysBig === 0n) return null;
-    const scannedEstimate = BigInt(Math.round(hashrateNum * elapsedSeconds));
-    const pct = Number(scannedEstimate * 10000n / jobKeysBig) / 100;
-    return Math.min(100, Math.max(0, pct));
-}
-
-// --- Legacy sector seeding retained for rollback compatibility ---
-function seedSectors(db, puzzleId, startHex, endHex) {
-    const MIN_SECTOR_SIZE = 1000000000n;
-
-    const start = BigInt('0x' + startHex);
-    const end   = BigInt('0x' + endHex);
-    const range = end - start;
-    if (range <= 0n) throw new Error('Puzzle range must be > 0');
-
-    let numSectors = range / MIN_SECTOR_SIZE;
-    if (numSectors < 1n)             numSectors = 1n;
-    if (numSectors > TARGET_SECTORS) numSectors = TARGET_SECTORS;
-
-    const sectorSize = range / numSectors;
-    const stmt = db.prepare(
-        "INSERT INTO sectors (puzzle_id, start_hex, end_hex, current_hex, status) VALUES (?, ?, ?, ?, 'open')"
-    );
-
-    db.transaction(() => {
-        for (let i = 0n; i < numSectors; i++) {
-            const s = start + i * sectorSize;
-            const e = (i === numSectors - 1n) ? end : s + sectorSize;
-            const sHex = s.toString(16).padStart(64, '0');
-            const eHex = e.toString(16).padStart(64, '0');
-            stmt.run(puzzleId, sHex, eHex, sHex);
-        }
-    })();
-
-    console.log(`[System] Seeded ${numSectors} legacy sectors for puzzle ${puzzleId}`);
-}
-
-// --- New allocator seeding: virtual chunks ---
-function seedVirtualChunks(db, puzzleId, startHex, endHex, allocSeed, virtualChunkSizeKeys = null) {
-    const { range } = normalizedRange(startHex, endHex);
-
-    const shardSize = bigIntMin(chooseShardSize(range), range);
-    const shardCountBig = ceilDiv(range, shardSize);
-    validateShardCountOrThrow(shardCountBig, 'seedVirtualChunks');
-    const shardCount = Number(shardCountBig);
-
-    let chunkSize = virtualChunkSizeKeys && virtualChunkSizeKeys > 0n
-        ? bigIntMin(virtualChunkSizeKeys, shardSize)
-        : chooseVirtualChunkSizeForShard(shardSize);
-
-    const vchunksPerShardBig = ceilDiv(shardSize, chunkSize);
-    const vchunksPerShardText = vchunksPerShardBig.toString();
-
-    db.transaction(() => {
-        db.prepare(`
-            UPDATE puzzles
-            SET alloc_strategy = ?,
-                alloc_seed = ?,
-                alloc_cursor = 0,
-                bootstrap_stage = 0,
-                shard_size_keys = ?,
-                shard_count = ?,
-                shard_alloc_cursor = 0,
-                shard_alloc_cursor_text = ?,
-                virtual_chunk_size_keys = ?,
-                virtual_chunk_count = ?,
-                virtual_chunk_count_text = ?,
-                virtual_chunks_per_shard = ?,
-                virtual_chunks_per_shard_text = ?
-            WHERE id = ?
-        `).run(
-            ALLOC_STRATEGY_VCHUNKS,
-            allocSeed,
-            shardSize.toString(),
-            shardCount,
-            encodeBigIndex(0n),
-            chunkSize.toString(),
-            vchunksPerShardText,
-            vchunksPerShardText,
-            vchunksPerShardText,
-            vchunksPerShardText,
-            puzzleId
-        );
-
-        db.prepare("DELETE FROM puzzle_shards WHERE puzzle_id = ?").run(puzzleId);
-
-        const insertShard = db.prepare(`
-            INSERT INTO puzzle_shards (
-                puzzle_id, shard_index, status, alloc_cursor_text, bootstrap_stage, assigned_runs, completed_runs
-            ) VALUES (?, ?, 'open', ?, 0, 0, 0)
-        `);
-
-        for (let i = 0; i < shardCount; i++) {
-            insertShard.run(puzzleId, i, encodeBigIndex(0n));
-        }
-    })();
-
-    console.log(
-        `[System] Seeded ${shardCount} shards for puzzle ${puzzleId} ` +
-        `(shard size ${shardSize.toString()} keys, ` +
-        `virtual chunk size ${chunkSize.toString()} keys, ` +
-        `${vchunksPerShardBig.toString()} vchunks/shard)`
-    );
-}
-
-function ensureAllocatorForPuzzle(db, puzzleId) {
-    const puzzle = db.prepare("SELECT * FROM puzzles WHERE id = ?").get(puzzleId);
-    if (!puzzle) throw new Error(`Puzzle ${puzzleId} not found`);
-
-    const strategy = puzzle.alloc_strategy || ALLOC_STRATEGY_LEGACY;
-
-    if (strategy === ALLOC_STRATEGY_VCHUNKS) {
-        const issuedCount = db.prepare(
-            "SELECT COUNT(*) AS c FROM chunks WHERE puzzle_id = ? AND is_test = 0"
-        ).get(puzzleId).c;
-
-        const seed = puzzle.alloc_seed || defaultAllocSeedForPuzzle(puzzle, ALLOC_STRATEGY_VCHUNKS);
-        const { range } = normalizedRange(puzzle.start_hex, puzzle.end_hex);
-
-        const desiredShardSize = bigIntMin(chooseShardSize(range), range);
-        const desiredChunkSize = (() => {
-            if (puzzle.virtual_chunk_size_keys) {
-                try {
-                    const stored = BigInt(puzzle.virtual_chunk_size_keys);
-                    if (stored > 0n) return bigIntMin(stored, desiredShardSize);
-                } catch (_) {}
-            }
-            return chooseVirtualChunkSizeForShard(desiredShardSize);
-        })();
-
-        const expectedShardCount = ceilDiv(range, desiredShardSize);
-        const expectedVChunksPerShard = ceilDiv(desiredShardSize, desiredChunkSize);
-
-        const storedShardCount = parsePositiveBigInt(puzzle.shard_count, null);
-        const storedShardSize = parsePositiveBigInt(puzzle.shard_size_keys, null);
-        const storedVChunkCount = decodeBigIndex(
-            puzzle.virtual_chunk_count_text,
-            parsePositiveBigInt(puzzle.virtual_chunk_count, null)
-        );
-        const storedPerShard = decodeBigIndex(
-            puzzle.virtual_chunks_per_shard_text,
-            parsePositiveBigInt(puzzle.virtual_chunks_per_shard, null)
-        );
-
-        const metadataLooksWrong =
-            !storedShardCount ||
-            !storedShardSize ||
-            !storedVChunkCount ||
-            !storedPerShard ||
-            storedShardCount !== expectedShardCount ||
-            storedShardSize !== desiredShardSize ||
-            storedVChunkCount !== expectedVChunksPerShard ||
-            storedPerShard !== expectedVChunksPerShard;
-
-        if (metadataLooksWrong) {
-            seedVirtualChunks(db, puzzleId, puzzle.start_hex, puzzle.end_hex, seed, desiredChunkSize);
-        } else if (
-            AUTO_RESEED_EMPTY_VCHUNK_PUZZLES &&
-            issuedCount === 0 &&
-            BigInt(puzzle.virtual_chunk_size_keys || desiredChunkSize.toString()) !== desiredChunkSize
-        ) {
-            console.log(
-                `[System] Re-seeding empty virtual-chunk puzzle ${puzzleId} ` +
-                `to shard-aware layout with chunk size ${desiredChunkSize.toString()}`
-            );
-            seedVirtualChunks(db, puzzleId, puzzle.start_hex, puzzle.end_hex, seed, desiredChunkSize);
-        }
-    } 
-    else {
-        const sectorCount = db.prepare("SELECT COUNT(*) AS c FROM sectors WHERE puzzle_id = ?").get(puzzleId).c;
-        if (sectorCount === 0) {
-            seedSectors(db, puzzle.id, puzzle.start_hex, puzzle.end_hex);
-        }
-    }
-}
-
-// --- App factory ---
-function createApp(db) {
-    const stmtWorkerHash = db.prepare("SELECT hashrate FROM workers WHERE name = ?");
-    const stmtChunkCount = db.prepare("SELECT COUNT(*) as cnt FROM chunks WHERE puzzle_id = ? AND is_test = 0");
-
-    // Legacy allocator statements
-    const stmtOpenSector    = db.prepare("SELECT * FROM sectors WHERE puzzle_id = ? AND status = 'open' ORDER BY RANDOM() LIMIT 1");
-    const stmtOpenSectorAt  = db.prepare("SELECT * FROM sectors WHERE puzzle_id = ? AND status = 'open' ORDER BY id ASC LIMIT 1 OFFSET ?");
-    const stmtSectorDone    = db.prepare("UPDATE sectors SET current_hex = end_hex, status = 'done' WHERE id = ?");
-    const stmtSectorAdvance = db.prepare("UPDATE sectors SET current_hex = ? WHERE id = ?");
-
-    // Virtual-chunk allocator statements
-    const stmtVChunkPuzzle = db.prepare(`
-        SELECT id, name, start_hex, end_hex,
-               alloc_strategy, alloc_seed, alloc_cursor,
-               shard_size_keys, shard_count, shard_alloc_cursor,
-               shard_alloc_cursor_text,
-               virtual_chunk_size_keys, virtual_chunk_count, virtual_chunk_count_text,
-               virtual_chunks_per_shard, virtual_chunks_per_shard_text,
-               bootstrap_stage,
-               test_start_hex, test_end_hex
-        FROM puzzles
-        WHERE id = ?
-    `);
-
-    const stmtShardByIndex = db.prepare(`
-        SELECT puzzle_id, shard_index, status, alloc_cursor_text, bootstrap_stage, assigned_runs, completed_runs
-        FROM puzzle_shards
-        WHERE puzzle_id = ? AND shard_index = ?
-    `);
-
-    const stmtShardCursorSet = db.prepare(`
-        UPDATE puzzle_shards
-        SET alloc_cursor_text = ?
-        WHERE puzzle_id = ? AND shard_index = ?
-    `);
-
-    const stmtShardBootstrapStageSet = db.prepare(`
-        UPDATE puzzle_shards
-        SET bootstrap_stage = ?
-        WHERE puzzle_id = ? AND shard_index = ?
-    `);
-
-    const stmtShardAssignedRunInc = db.prepare(`
-        UPDATE puzzle_shards
-        SET assigned_runs = assigned_runs + 1
-        WHERE puzzle_id = ? AND shard_index = ?
-    `);
-
-    const stmtShardCompletedRunInc = db.prepare(`
-        UPDATE puzzle_shards
-        SET completed_runs = completed_runs + 1
-        WHERE puzzle_id = ? AND shard_index = ?
-    `);
-
-    const stmtShardLocalAnyOverlap = db.prepare(`
-        SELECT 1
-        FROM chunks
-        WHERE puzzle_id = ?
-          AND shard_index = ?
-          AND is_test = 0
-          AND status IN ('assigned', 'reclaimed', 'completed', 'FOUND')
-          AND local_vchunk_start_text IS NOT NULL
-          AND local_vchunk_end_text IS NOT NULL
-          AND local_vchunk_start_text < ?
-          AND local_vchunk_end_text > ?
-        LIMIT 1
-    `);
-
-    const stmtInsertChunk = db.prepare(`
-        INSERT INTO chunks (
-            puzzle_id, start_hex, end_hex, status,
-            worker_name, assigned_at, heartbeat_at, is_test,
-            sector_id, alloc_block_id,
-            vchunk_start, vchunk_end,
-            alloc_generation,
-            shard_index, local_vchunk_start_text, local_vchunk_end_text
-        ) VALUES (?, ?, ?, 'assigned', ?, CURRENT_TIMESTAMP, CURRENT_TIMESTAMP, 0, NULL, NULL, NULL, NULL, ?, ?, ?, ?)
-    `);
-
-    const stmtTestChunkTaken = db.prepare(`
-        SELECT id
-        FROM chunks
-        WHERE puzzle_id = ? AND start_hex = ? AND end_hex = ? AND is_test = 1 AND status = 'assigned'
-        LIMIT 1
-    `);
-
-    const stmtTestChunkReclaim = db.prepare(`
-        UPDATE chunks
-        SET status = 'assigned',
-            worker_name = ?,
-            assigned_at = CURRENT_TIMESTAMP,
-            heartbeat_at = CURRENT_TIMESTAMP
-        WHERE id = (
-            SELECT id
-            FROM chunks
-            WHERE puzzle_id = ? AND start_hex = ? AND end_hex = ? AND is_test = 1 AND status = 'reclaimed'
-            LIMIT 1
-        )
-        RETURNING *
-    `);
-
-    const stmtTestChunkInsert = db.prepare(`
-        INSERT INTO chunks (
-            puzzle_id, start_hex, end_hex, status, worker_name,
-            assigned_at, heartbeat_at, is_test, alloc_generation
-        )
-        VALUES (?, ?, ?, 'assigned', ?, CURRENT_TIMESTAMP, CURRENT_TIMESTAMP, 1, 'test')
-        RETURNING *
-    `);
-
-    const stmtPuzzleShardCursorTextSet = db.prepare(`
-        UPDATE puzzles
-        SET shard_alloc_cursor_text = ?
-        WHERE id = ?
-    `);
-
-    const MIDPOINT_SECTOR = Number(TARGET_SECTORS / 2n);
-
-    const assignLegacyRandomChunk = db.transaction((name, hashrate, puzzle) => {
-        const hashrateBig = normalizeHashrate(hashrate || stmtWorkerHash.get(name)?.hashrate);
-        const chunkSize   = hashrateBig * BigInt(TARGET_MINUTES * 60);
-        const isFirst = stmtChunkCount.get(puzzle.id).cnt === 0;
-
-        for (;;) {
-            const sector = isFirst
-                ? (stmtOpenSectorAt.get(puzzle.id, MIDPOINT_SECTOR) || stmtOpenSector.get(puzzle.id))
-                : stmtOpenSector.get(puzzle.id);
-            if (!sector) return null;
-
-            const current   = BigInt('0x' + sector.current_hex);
-            const sectorEnd = BigInt('0x' + sector.end_hex);
-            const effective = bigIntMin(chunkSize, sectorEnd - current);
-
-            if (effective <= 0n) {
-                stmtSectorDone.run(sector.id);
-                continue;
-            }
-
-            const startHex = current.toString(16).padStart(64, '0');
-            const endHex   = (current + effective).toString(16).padStart(64, '0');
-
-            if (current + effective >= sectorEnd) {
-                stmtSectorDone.run(sector.id);
-            } else {
-                stmtSectorAdvance.run(endHex, sector.id);
-            }
-
-            const info = db.prepare(`
-                INSERT INTO chunks (
-                    puzzle_id, start_hex, end_hex, status,
-                    worker_name, assigned_at, heartbeat_at, is_test,
-                    sector_id, alloc_block_id, vchunk_start, vchunk_end,
-                    alloc_generation
-                ) VALUES (?, ?, ?, 'assigned', ?, CURRENT_TIMESTAMP, CURRENT_TIMESTAMP, 0, ?, NULL, NULL, NULL, 'legacy')
-            `).run(puzzle.id, startHex, endHex, name, sector.id);
-
-            return { chunkId: info.lastInsertRowid, startHex, endHex };
-        }
-    });
-
-    function rangeIsFreeInShard(puzzleId, shardIndex, localStartBig, localEndExclusiveBig) {
-        if (localStartBig < 0n || localEndExclusiveBig <= localStartBig) return false;
-
-        return !stmtShardLocalAnyOverlap.get(
-            puzzleId,
-            shardIndex,
-            encodeBigIndex(localEndExclusiveBig),
-            encodeBigIndex(localStartBig)
-        );
-    }
-
-    function findBeginBootstrapRun(puzzleId, shardIndex, totalChunksBig, neededChunksBig) {
-        const maxStart = totalChunksBig - neededChunksBig;
-        if (maxStart < 0n) return null;
-
-        const probeLimitBig = bigIntMin(maxStart + 1n, BigInt(MAX_ALLOC_PROBES));
-        const probeLimit = Number(probeLimitBig);
-
-        for (let i = 0; i < probeLimit; i++) {
-            const start = BigInt(i);
-            if (rangeIsFreeInShard(puzzleId, shardIndex, start, start + neededChunksBig)) return start;
-        }
-
-        return null;
-    }
-
-    function findEndBootstrapRun(puzzleId, shardIndex, totalChunksBig, neededChunksBig) {
-        const maxStart = totalChunksBig - neededChunksBig;
-        if (maxStart < 0n) return null;
-
-        const probeLimitBig = bigIntMin(maxStart + 1n, BigInt(MAX_ALLOC_PROBES));
-        const probeLimit = Number(probeLimitBig);
-
-        for (let i = 0; i < probeLimit; i++) {
-            const start = maxStart - BigInt(i);
-            if (rangeIsFreeInShard(puzzleId, shardIndex, start, start + neededChunksBig)) return start;
-        }
-
-        return null;
-    }
-
-    function findMidBootstrapRun(puzzleId, shardIndex, totalChunksBig, neededChunksBig) {
-        if (totalChunksBig <= 0n) return null;
-
-        const anchor = totalChunksBig / 2n;
-        const maxStart = totalChunksBig - neededChunksBig;
-        if (maxStart < 0n) return null;
-
-        const tried = new Set();
-
-        for (let dist = 0; dist < MAX_ALLOC_PROBES; dist++) {
-            const distBig = BigInt(dist);
-
-            const leftAnchor = anchor - distBig;
-            if (leftAnchor >= 0n) {
-                const start = normalizeRunStartForCandidateBig(leftAnchor, neededChunksBig, totalChunksBig);
-                const k = encodeBigIndex(start);
-                if (!tried.has(k)) {
-                    tried.add(k);
-                    if (rangeIsFreeInShard(puzzleId, shardIndex, start, start + neededChunksBig)) return start;
-                }
-            }
-
-            if (dist === 0) continue;
-
-            const rightAnchor = anchor + distBig;
-            if (rightAnchor < totalChunksBig) {
-                const start = normalizeRunStartForCandidateBig(rightAnchor, neededChunksBig, totalChunksBig);
-                const k = encodeBigIndex(start);
-                if (!tried.has(k)) {
-                    tried.add(k);
-                    if (rangeIsFreeInShard(puzzleId, shardIndex, start, start + neededChunksBig)) return start;
-                }
-            }
-        }
-
-        return null;
-    }
-
-    function assignVirtualChunkRun(name, puzzle, shardIndex, localRunStartBig, runCountBig) {
-        const localRunEndBig = localRunStartBig + runCountBig;
-
-        const { startHex, endHex } = shardLocalVirtualChunkRangeToHex(
-            puzzle,
-            shardIndex,
-            localRunStartBig,
-            localRunEndBig
-        );
-
-        const generation = PERMUTATION_MODE;
-
-        const info = stmtInsertChunk.run(
-            puzzle.id,
-            startHex,
-            endHex,
-            name,
-            generation,
-            shardIndex,
-            encodeBigIndex(localRunStartBig),
-            encodeBigIndex(localRunEndBig)
-        );
-
-        stmtShardAssignedRunInc.run(puzzle.id, shardIndex);
-
-        return {
-            chunkId: info.lastInsertRowid,
-            startHex,
-            endHex,
-            shardIndex,
-            localVChunkStart: localRunStartBig,
-            localVChunkEnd: localRunEndBig
-        };
-    }
-
-    const assignVirtualChunkJob = db.transaction((name, workHints, puzzle) => {
-        const currentPuzzle = stmtVChunkPuzzle.get(puzzle.id);
-        if (!currentPuzzle) return null;
-        const perShardChunksBig = decodeBigIndex(
-            currentPuzzle.virtual_chunks_per_shard_text,
-            decodeBigIndex(
-                currentPuzzle.virtual_chunks_per_shard,
-                decodeBigIndex(
-                    currentPuzzle.virtual_chunk_count_text,
-                    decodeBigIndex(currentPuzzle.virtual_chunk_count, null)
-                )
-            )
-        );
-
-        if (!currentPuzzle.virtual_chunk_size_keys || !perShardChunksBig) return null;
-        if (!currentPuzzle.shard_count || !currentPuzzle.shard_size_keys) return null;
-
-        const shardCountBig = BigInt(currentPuzzle.shard_count);
-        if (shardCountBig <= 0n) return null;
-
-        const totalShards = Number(shardCountBig);
-        if (perShardChunksBig <= 0n) return null;
-
-        const hashrateBig = normalizeHashrate(workHints.hashrate || stmtWorkerHash.get(name)?.hashrate);
-        const minChunkKeys = parsePositiveBigInt(workHints.min_chunk_keys, null);
-        const chunkQuantumKeys = parsePositiveBigInt(workHints.chunk_quantum_keys, 1n);
-        const requestedKeys = computeWorkerRequestedKeys({ hashrateBig, minChunkKeys, chunkQuantumKeys });
-
-        const vchunkSize = BigInt(currentPuzzle.virtual_chunk_size_keys);
-        let neededChunksBig = ceilDiv(requestedKeys, vchunkSize);
-        if (neededChunksBig < 1n) neededChunksBig = 1n;
-        if (neededChunksBig > perShardChunksBig) neededChunksBig = perShardChunksBig;
-
-        console.log(
-            `[Alloc] ${name} requested ${requestedKeys.toString()} keys ` +
-            `(hashrate=${hashrateBig.toString()}, min=${minChunkKeys ? minChunkKeys.toString() : 'null'}, ` +
-            `quantum=${chunkQuantumKeys ? chunkQuantumKeys.toString() : '1'}) ` +
-            `=> need ${neededChunksBig.toString()} virtual chunks of ${vchunkSize.toString()} keys ` +
-            `(shards=${totalShards}, perShard=${perShardChunksBig.toString()})`
-        );
-
-        const permKey = currentPuzzle.alloc_seed || defaultAllocSeedForPuzzle(currentPuzzle, ALLOC_STRATEGY_VCHUNKS);
-        const rawShardCursorBig = decodeBigIndex(
-            currentPuzzle.shard_alloc_cursor_text,
-            decodeBigIndex(currentPuzzle.shard_alloc_cursor, 0n)
-        );
-        const shardProbeLimit = Math.min(totalShards, MAX_ALLOC_PROBES);
-
-        for (let shardOffset = 0; shardOffset < shardProbeLimit; shardOffset++) {
-            const shardOrderIndex = (rawShardCursorBig + BigInt(shardOffset)) % shardCountBig;
-            const shardIndex = pickShardIndex(shardOrderIndex, shardCountBig, permKey);
-
-            const shard = stmtShardByIndex.get(currentPuzzle.id, shardIndex);
-            if (!shard || shard.status !== 'open') continue;
-
-            const stage = Number(shard.bootstrap_stage || 0);
-            const freshRuns = Number(shard.assigned_runs || 0);
-            const rawCursorBig = decodeBigIndex(shard.alloc_cursor_text, 0n);
-
-            if (freshRuns < 3 && stage < 3) {
-                let runStartBig = null;
-
-                if (stage === 0) runStartBig = findMidBootstrapRun(currentPuzzle.id, shardIndex, perShardChunksBig, neededChunksBig);
-                else if (stage === 1) runStartBig = findBeginBootstrapRun(currentPuzzle.id, shardIndex, perShardChunksBig, neededChunksBig);
-                else if (stage === 2) runStartBig = findEndBootstrapRun(currentPuzzle.id, shardIndex, perShardChunksBig, neededChunksBig);
-
-                if (runStartBig !== null) {
-                    const result = assignVirtualChunkRun(name, currentPuzzle, shardIndex, runStartBig, neededChunksBig);
-
-                    stmtShardBootstrapStageSet.run(stage + 1, currentPuzzle.id, shardIndex);
-                    stmtShardCursorSet.run(
-                        encodeBigIndex((runStartBig + 1n) % perShardChunksBig),
-                        currentPuzzle.id,
-                        shardIndex
-                    );
-                    stmtPuzzleShardCursorTextSet.run(
-                        encodeBigIndex((shardOrderIndex + 1n) % shardCountBig),
-                        currentPuzzle.id
-                    );
-
-                    console.log(
-                        `[Alloc] bootstrap shard ${shardIndex} stage ${stage} -> assigned ${name} ` +
-                        `local vchunks ${result.localVChunkStart.toString()}..${(result.localVChunkEnd - 1n).toString()} ` +
-                        `(${result.startHex} .. ${result.endHex})`
-                    );
-                    return result;
-                }
-
-                for (let fallback = neededChunksBig - 1n; fallback >= 1n; fallback--) {
-                    if (stage === 0) runStartBig = findMidBootstrapRun(currentPuzzle.id, shardIndex, perShardChunksBig, fallback);
-                    else if (stage === 1) runStartBig = findBeginBootstrapRun(currentPuzzle.id, shardIndex, perShardChunksBig, fallback);
-                    else if (stage === 2) runStartBig = findEndBootstrapRun(currentPuzzle.id, shardIndex, perShardChunksBig, fallback);
-
-                    if (runStartBig !== null) {
-                        const result = assignVirtualChunkRun(name, currentPuzzle, shardIndex, runStartBig, fallback);
-
-                        stmtShardBootstrapStageSet.run(stage + 1, currentPuzzle.id, shardIndex);
-                        stmtShardCursorSet.run(
-                            encodeBigIndex((runStartBig + 1n) % perShardChunksBig),
-                            currentPuzzle.id,
-                            shardIndex
-                        );
-                        stmtPuzzleShardCursorTextSet.run(
-                            encodeBigIndex((shardOrderIndex + 1n) % shardCountBig),
-                            currentPuzzle.id
-                        );
-
-                        console.log(
-                            `[Alloc] bootstrap fallback shard ${shardIndex} stage ${stage} -> assigned ${name} ` +
-                            `local vchunks ${result.localVChunkStart.toString()}..${(result.localVChunkEnd - 1n).toString()} ` +
-                            `(${result.startHex} .. ${result.endHex})`
-                        );
-                        return result;
-                    }
-                }
-
-                stmtShardBootstrapStageSet.run(stage + 1, currentPuzzle.id, shardIndex);
-            }
-
-            let a = null;
-            let b = null;
-            if (PERMUTATION_MODE === 'affine') {
-                const affine = deriveAffinePermutationParams(`${permKey}:shard:${shardIndex}`, perShardChunksBig);
-                a = affine.a;
-                b = affine.b;
-            }
-
-            const triedStarts = new Set();
-
-            for (let offset = 0; offset < MAX_ALLOC_PROBES; offset++) {
-                const orderIndex = (rawCursorBig + BigInt(offset)) % perShardChunksBig;
-                const candidateIndex = (
-                    PERMUTATION_MODE === 'affine'
-                        ? permuteIndexAffine(orderIndex, perShardChunksBig, a, b)
-                        : permuteIndexFeistel(orderIndex, perShardChunksBig, `${permKey}:shard:${shardIndex}`)
-                );
-
-                const runStartBig = normalizeRunStartForCandidateBig(candidateIndex, neededChunksBig, perShardChunksBig);
-                const key = encodeBigIndex(runStartBig);
-                if (triedStarts.has(key)) continue;
-                triedStarts.add(key);
-
-                if (rangeIsFreeInShard(currentPuzzle.id, shardIndex, runStartBig, runStartBig + neededChunksBig)) {
-                    stmtShardCursorSet.run(
-                        encodeBigIndex((orderIndex + 1n) % perShardChunksBig),
-                        currentPuzzle.id,
-                        shardIndex
-                    );
-                    stmtPuzzleShardCursorTextSet.run(
-                        encodeBigIndex((shardOrderIndex + 1n) % shardCountBig),
-                        currentPuzzle.id
-                    );
-
-                    const result = assignVirtualChunkRun(name, currentPuzzle, shardIndex, runStartBig, neededChunksBig);
-                    if (stage < 3) stmtShardBootstrapStageSet.run(3, currentPuzzle.id, shardIndex);
-
-                    console.log(
-                        `[Alloc] ${PERMUTATION_MODE} shard ${shardIndex} full-run -> assigned ${name} ` +
-                        `local vchunks ${result.localVChunkStart.toString()}..${(result.localVChunkEnd - 1n).toString()} ` +
-                        `(${result.startHex} .. ${result.endHex})`
-                    );
-                    return result;
-                }
-            }
-
-            for (let fallback = neededChunksBig - 1n; fallback >= 1n; fallback--) {
-                const triedFallback = new Set();
-
-                for (let offset = 0; offset < MAX_ALLOC_PROBES; offset++) {
-                    const orderIndex = (rawCursorBig + BigInt(offset)) % perShardChunksBig;
-                    const candidateIndex = (
-                        PERMUTATION_MODE === 'affine'
-                            ? permuteIndexAffine(orderIndex, perShardChunksBig, a, b)
-                            : permuteIndexFeistel(orderIndex, perShardChunksBig, `${permKey}:shard:${shardIndex}`)
-                    );
-
-                    const runStartBig = normalizeRunStartForCandidateBig(candidateIndex, fallback, perShardChunksBig);
-                    const key = encodeBigIndex(runStartBig);
-                    if (triedFallback.has(key)) continue;
-                    triedFallback.add(key);
-
-                    if (rangeIsFreeInShard(currentPuzzle.id, shardIndex, runStartBig, runStartBig + fallback)) {
-                        stmtShardCursorSet.run(
-                            encodeBigIndex((orderIndex + 1n) % perShardChunksBig),
-                            currentPuzzle.id,
-                            shardIndex
-                        );
-                        stmtPuzzleShardCursorTextSet.run(
-                            encodeBigIndex((shardOrderIndex + 1n) % shardCountBig),
-                            currentPuzzle.id
-                        );
-
-                        const result = assignVirtualChunkRun(name, currentPuzzle, shardIndex, runStartBig, fallback);
-                        if (stage < 3) stmtShardBootstrapStageSet.run(3, currentPuzzle.id, shardIndex);
-
-                        console.log(
-                            `[Alloc] ${PERMUTATION_MODE} shard ${shardIndex} fallback -> assigned ${name} ` +
-                            `local vchunks ${result.localVChunkStart.toString()}..${(result.localVChunkEnd - 1n).toString()} ` +
-                            `(${result.startHex} .. ${result.endHex})`
-                        );
-                        return result;
-                    }
-                }
-            }
-        }
-
-        console.log(`[Alloc] no free shard-local run found for ${name}`);
-        return null;
-    });
-
-    const claimTestChunk = db.transaction((name, puzzle) => {
-        const taken = stmtTestChunkTaken.get(puzzle.id, puzzle.test_start_hex, puzzle.test_end_hex);
-        if (taken) return null;
-        const reclaimed = stmtTestChunkReclaim.get(name, puzzle.id, puzzle.test_start_hex, puzzle.test_end_hex);
-        if (reclaimed) return reclaimed;
-        return stmtTestChunkInsert.get(puzzle.id, puzzle.test_start_hex, puzzle.test_end_hex, name);
-    });
-
-    const app = express();
-    app.use(express.json());
-
-    app.get('/', (req, res) => {
-        res.set('Cache-Control', 'no-store');
-        res.sendFile('public/index.html', { root: '.' });
-    });
-    app.use(express.static('public'));
-
-    if (process.env.ADMIN_TOKEN) {
-        app.use('/api/v1/admin', (req, res, next) => {
-            if (req.headers['x-admin-token'] === process.env.ADMIN_TOKEN) return next();
-            return res.status(401).json({ error: 'unauthorized' });
-        });
-    }
-
-    // --- API Endpoints ---
-
-    app.post('/api/v1/work', (req, res) => {
-        const { name, hashrate, version, min_chunk_keys, chunk_quantum_keys } = req.body;
-        if (!name) return res.status(400).json({ error: 'Missing name' });
-
-        const hashrateNum = Number(normalizeHashrate(hashrate));
-        const minChunkKeys = parsePositiveBigInt(min_chunk_keys, null);
-        const chunkQuantumKeys = parsePositiveBigInt(chunk_quantum_keys, null);
-
-        const prevWorker = db.prepare(
-            `SELECT CASE WHEN last_seen < datetime('now', '-${REACTIVATE_MINUTES} minutes') THEN 1 ELSE 0 END AS inactive FROM workers WHERE name = ?`
-        ).get(name);
-        const isReactivating = prevWorker?.inactive === 1;
-
-        if (isReactivating) {
-            db.prepare(`
-                UPDATE chunks
-                SET status = 'reclaimed', prev_worker_name = worker_name, worker_name = NULL, assigned_at = NULL, heartbeat_at = NULL
-                WHERE worker_name = ? AND status = 'assigned'
-            `).run(name);
-        }
-
-        db.prepare(`
-            INSERT INTO workers (name, hashrate, last_seen, version, min_chunk_keys, chunk_quantum_keys)
-            VALUES (?, ?, CURRENT_TIMESTAMP, ?, ?, ?)
-            ON CONFLICT(name) DO UPDATE
-            SET hashrate = excluded.hashrate,
-                last_seen = CURRENT_TIMESTAMP,
-                version = COALESCE(excluded.version, workers.version),
-                min_chunk_keys = COALESCE(excluded.min_chunk_keys, workers.min_chunk_keys),
-                chunk_quantum_keys = COALESCE(excluded.chunk_quantum_keys, workers.chunk_quantum_keys)
-        `).run(
-            name,
-            hashrateNum,
-            version || null,
-            minChunkKeys ? minChunkKeys.toString() : null,
-            chunkQuantumKeys ? chunkQuantumKeys.toString() : null
-        );
-
-        const puzzle = db.prepare("SELECT * FROM puzzles WHERE active = 1 LIMIT 1").get();
-        if (!puzzle) return res.status(503).json({ error: 'No active puzzle configured' });
-
-        ensureAllocatorForPuzzle(db, puzzle.id);
-
-        const existing = db.prepare(`
-            SELECT id, start_hex, end_hex
-            FROM chunks
-            WHERE worker_name = ? AND puzzle_id = ? AND status = 'assigned'
-            LIMIT 1
-        `).get(name, puzzle.id);
-        if (existing) {
-            // Strict liveness model: only POST /heartbeat resets the reclaim timer.
-            // A repeated /work call for an already-assigned chunk re-issues the job_id
-            // but does NOT update heartbeat_at. Workers must call /heartbeat to stay alive.
-            return res.json({ job_id: existing.id, start_key: existing.start_hex, end_key: existing.end_hex });
-        }
-
-        if (puzzle.test_start_hex && puzzle.test_end_hex) {
-            const claimed = claimTestChunk.immediate(name, puzzle);
-            if (claimed) {
-                console.log(`[Test] Assigned test chunk #${claimed.id} to ${name}`);
-                return res.json({ job_id: claimed.id, start_key: claimed.start_hex, end_key: claimed.end_hex });
-            }
-        }
-
-        const reclaimed = isReactivating ? null : db.prepare(`
-            UPDATE chunks
-            SET status = 'assigned',
-                worker_name = ?,
-                assigned_at = CURRENT_TIMESTAMP,
-                heartbeat_at = CURRENT_TIMESTAMP
-            WHERE id = (
-                SELECT id
-                FROM chunks
-                WHERE status = 'reclaimed' AND puzzle_id = ? AND is_test = 0
-                ORDER BY id ASC LIMIT 1
-            )
-            RETURNING *
-        `).get(name, puzzle.id);
-
-        let chunkId, startHex, endHex;
-
-        if (reclaimed) {
-            chunkId  = reclaimed.id;
-            startHex = reclaimed.start_hex;
-            endHex   = reclaimed.end_hex;
-        } else {
-            let result = null;
-            const strategy = puzzle.alloc_strategy || ALLOC_STRATEGY_LEGACY;
-
-            if (strategy === ALLOC_STRATEGY_VCHUNKS) {
-                result = assignVirtualChunkJob.immediate(name, {
-                    hashrate,
-                    min_chunk_keys,
-                    chunk_quantum_keys
-                }, puzzle);
-            } else {
-                result = assignLegacyRandomChunk.immediate(name, hashrate, puzzle);
-            }
-
-            if (!result) return res.status(503).json({ error: 'All keyspace has been assigned' });
-            ({ chunkId, startHex, endHex } = result);
-        }
-
-        res.json({ job_id: chunkId, start_key: startHex, end_key: endHex });
-    });
-
-    app.post('/api/v1/submit', (req, res) => {
-        const { name, job_id, status, findings } = req.body;
-        if (status !== 'done' && status !== 'FOUND') {
-            return res.status(400).json({ error: 'status must be "done" or "FOUND"' });
-        }
-
-        if (status === 'FOUND') {
-            if (!Array.isArray(findings) || findings.length === 0) {
-                return res.status(400).json({ error: 'findings must be a non-empty array' });
-            }
-            for (const f of findings) {
-                if (!f || typeof f !== 'object' || Array.isArray(f)) {
-                    return res.status(400).json({ error: 'each finding must be a plain object' });
-                }
-                if (!f.found_key) {
-                    return res.status(400).json({ error: 'each finding must include found_key' });
-                }
-                if (!isValidHex(f.found_key)) {
-                    return res.status(400).json({ error: 'each found_key must be a valid hex string' });
-                }
-            }
-
-            const normalized = findings.map(f => ({ ...f, found_key: normalizeHex(f.found_key) }));
-            const seen = new Set();
-            const allFindings = normalized.filter(f => seen.has(f.found_key) ? false : (seen.add(f.found_key), true));
-            const primaryKey  = allFindings[0].found_key;
-            const primaryAddr = allFindings[0].found_address || null;
-
-            const stmtInsertFinding = db.prepare(`
-                INSERT OR IGNORE INTO findings (chunk_id, worker_name, found_key, found_address)
-                VALUES (?, ?, ?, ?)
-            `);
-
-            const result = db.transaction(() => {
-                const info = db.prepare(`
-                    UPDATE chunks
-                    SET status = 'FOUND',
-                        found_key = COALESCE(found_key, ?),
-                        found_address = COALESCE(found_address, ?)
-                    WHERE id = ? AND worker_name = ? AND status = 'assigned'
-                `).run(primaryKey, primaryAddr, job_id, name);
-
-                let isLate = false;
-                let completedNow = info.changes > 0;
-
-                if (!info.changes) {
-                    const alreadyRecorded = db.prepare(`
-                        SELECT id FROM findings WHERE chunk_id = ? AND worker_name = ? AND found_key = ?
-                    `).get(job_id, name, primaryKey);
-
-                    if (!alreadyRecorded) {
-                        const latePrev = db.prepare(`
-                            SELECT id, status
-                            FROM chunks
-                            WHERE id = ? AND prev_worker_name = ? AND status IN ('assigned', 'reclaimed')
-                        `).get(job_id, name);
-
-                        if (!latePrev) return { accepted: false, inserted: [], isLate: false, completedNow: false };
-
-                        db.prepare(`
-                            UPDATE chunks
-                            SET status = 'FOUND',
-                                worker_name = ?,
-                                found_key = COALESCE(found_key, ?),
-                                found_address = COALESCE(found_address, ?)
-                            WHERE id = ?
-                        `).run(name, primaryKey, primaryAddr, job_id);
-
-                        isLate = true;
-                        completedNow = true;
-                    }
-                }
-
-                const inserted = [];
-                for (const f of allFindings) {
-                    const r = stmtInsertFinding.run(job_id, name, f.found_key, f.found_address || null);
-                    if (r.changes) inserted.push(f);
-                }
-
-                return { accepted: true, inserted, isLate, completedNow };
-            })();
-
-            if (!result.accepted) return res.json({ accepted: false });
-            if (result.isLate) {
-                console.log(`[Late FOUND] Job: ${job_id} | Worker: ${name} (prev assignee) | KEY: ${primaryKey}`);
-            }
-
-            for (const f of result.inserted) {
-                const msg = `[${new Date().toISOString()}] BINGO! Job: ${job_id} | Worker: ${name} | KEY: ${f.found_key} | ADDR: ${f.found_address || 'Unknown'}\n`;
-                console.log(`\n🚨🚨🚨 ${msg}`);
-                fs.appendFileSync('BINGO_FOUND_KEYS.txt', msg);
-            }
-
-            if (result.completedNow) {
-                const completedChunk = db.prepare(`
-                    SELECT puzzle_id, shard_index
-                    FROM chunks
-                    WHERE id = ?
-                `).get(job_id);
-
-                if (completedChunk && completedChunk.shard_index !== null && completedChunk.shard_index !== undefined) {
-                    stmtShardCompletedRunInc.run(completedChunk.puzzle_id, completedChunk.shard_index);
-                }
-            }
-        } else {
-            const { keys_scanned } = req.body;
-
-            if (keys_scanned === undefined) {
-                return res.status(400).json({ accepted: false, error: 'keys_scanned is required for status: done' });
-            }
-
-            if (typeof keys_scanned !== 'number' || !Number.isInteger(keys_scanned) || keys_scanned < 0) {
-                return res.status(400).json({ accepted: false, error: 'keys_scanned must be a non-negative integer' });
-            }
-
-            const result = db.transaction(() => {
-                const chunk = db.prepare(`
-                    SELECT start_hex, end_hex
-                    FROM chunks
-                    WHERE id = ? AND worker_name = ? AND status = 'assigned'
-                `).get(job_id, name);
-                if (!chunk) return { notOwned: true };
-
-                const expectedSize = BigInt('0x' + chunk.end_hex) - BigInt('0x' + chunk.start_hex);
-                const reported = BigInt(keys_scanned);
-
-                if (reported < expectedSize) {
-                    const upd = db.prepare(`
-                        UPDATE chunks
-                        SET status = 'reclaimed', prev_worker_name = worker_name, worker_name = NULL, assigned_at = NULL, heartbeat_at = NULL
-                        WHERE id = ? AND worker_name = ? AND status = 'assigned'
-                    `).run(job_id, name);
-                    return { reclaimed: upd.changes > 0, expectedSize, reported };
-                }
-
-                return { sufficient: true };
-            })();
-
-            if (result.notOwned) return res.json({ accepted: false });
-            if (result.reclaimed) {
-                return res.status(400).json({
-                    accepted: false,
-                    error: `chunk #${job_id} not accepted, reported size: ${result.reported}, expected size: ${result.expectedSize}. Chunk reclaimed.`
-                });
-            }
-            if (!result.sufficient) return res.json({ accepted: false });
-
-            const info = db.prepare(`
-                UPDATE chunks SET status = 'completed'
-                WHERE id = ? AND worker_name = ? AND status = 'assigned'
-            `).run(job_id, name);
-            if (!info.changes) return res.json({ accepted: false });
-
-            const completedChunk = db.prepare(`
-                SELECT puzzle_id, shard_index
-                FROM chunks
-                WHERE id = ?
-            `).get(job_id);
-
-            if (completedChunk && completedChunk.shard_index !== null && completedChunk.shard_index !== undefined) {
-                stmtShardCompletedRunInc.run(completedChunk.puzzle_id, completedChunk.shard_index);
-            }
-        }
-
-        const chunk = db.prepare(`
-            SELECT is_test, puzzle_id, start_hex, end_hex
-            FROM chunks
-            WHERE id = ?
-        `).get(job_id);
-
-        if (chunk?.is_test) {
-            const cleared = db.prepare(`
-                UPDATE puzzles
-                SET test_start_hex = NULL, test_end_hex = NULL
-                WHERE id = ? AND test_start_hex = ? AND test_end_hex = ?
-            `).run(chunk.puzzle_id, chunk.start_hex, chunk.end_hex);
-
-            if (cleared.changes) {
-                console.log(`[Test] Test chunk #${job_id} completed — test config cleared from puzzle ${chunk.puzzle_id}`);
-            }
-        }
-
-        res.json({ accepted: true });
-    });
-
-    app.get('/api/v1/stats', (req, res) => {
-        const puzzleIdParam = req.query.puzzle_id ? parseInt(req.query.puzzle_id, 10) : null;
-        const puzzle = puzzleIdParam
-            ? (db.prepare("SELECT * FROM puzzles WHERE id = ?").get(puzzleIdParam) || null)
-            : (db.prepare("SELECT * FROM puzzles WHERE active = 1 LIMIT 1").get() || null);
-
-        const pid = puzzle ? puzzle.id : null;
-
-        const visibleWorkers = pid ? db.prepare(`
-            SELECT w.name, w.hashrate, w.last_seen, w.version, w.min_chunk_keys, w.chunk_quantum_keys,
-                   CASE WHEN w.last_seen >= datetime('now', '-${ACTIVE_MINUTES} minutes')
-                   THEN 1 ELSE 0 END AS fresh,
-                   CASE WHEN EXISTS (
-                       SELECT 1 FROM chunks c2
-                       WHERE c2.worker_name = w.name AND c2.puzzle_id = ? AND c2.status = 'assigned'
-                   ) THEN 1 ELSE 0 END AS assigned_here,
-                   CASE WHEN w.last_seen >= datetime('now', '-${ACTIVE_MINUTES} minutes')
-                        AND EXISTS (
-                            SELECT 1 FROM chunks c2
-                            WHERE c2.worker_name = w.name AND c2.puzzle_id = ? AND c2.status = 'assigned'
-                        )
-                   THEN 1 ELSE 0 END AS active
-            FROM workers w
-            WHERE w.last_seen >= datetime('now', '-${TIMEOUT_MINUTES} minutes')
-              AND EXISTS (
-                SELECT 1 FROM chunks c
-                WHERE (c.worker_name = w.name OR c.prev_worker_name = w.name)
-                  AND c.puzzle_id = ?
-              )
-            ORDER BY w.hashrate DESC
-        `).all(pid, pid, pid) : [];
-
-        const totalHashrate = visibleWorkers.filter(w => w.active).reduce((sum, w) => sum + w.hashrate, 0);
-
-        const completedChunks = pid ? db.prepare(`
-            SELECT COUNT(*) AS count
-            FROM chunks
-            WHERE puzzle_id = ? AND (status = 'completed' OR status = 'FOUND') AND is_test = 0
-        `).get(pid).count : 0;
-
-        const reclaimedChunks = pid ? db.prepare(`
-            SELECT COUNT(*) AS count
-            FROM chunks
-            WHERE puzzle_id = ? AND status = 'reclaimed' AND is_test = 0
-        `).get(pid).count : 0;
-
-        const doneChunks = pid ? db.prepare(`
-            SELECT worker_name, start_hex, end_hex
-            FROM chunks
-            WHERE puzzle_id = ? AND (status = 'completed' OR status = 'FOUND') AND worker_name IS NOT NULL AND is_test = 0
-        `).all(pid) : [];
-
-        let totalKeysCompleted = 0n;
-        if (doneChunks.length > 0) {
-            const sorted = [...doneChunks].sort((a, b) => {
-                const diff = BigInt('0x' + a.start_hex) - BigInt('0x' + b.start_hex);
-                return diff < 0n ? -1 : diff > 0n ? 1 : 0;
-            });
-            let mergeStart = BigInt('0x' + sorted[0].start_hex);
-            let mergeEnd   = BigInt('0x' + sorted[0].end_hex);
-            for (let i = 1; i < sorted.length; i++) {
-                const s = BigInt('0x' + sorted[i].start_hex);
-                const e = BigInt('0x' + sorted[i].end_hex);
-                if (s <= mergeEnd) {
-                    if (e > mergeEnd) mergeEnd = e;
-                } else {
-                    totalKeysCompleted += mergeEnd - mergeStart;
-                    mergeStart = s;
-                    mergeEnd   = e;
-                }
-            }
-            totalKeysCompleted += mergeEnd - mergeStart;
-        }
-
-        const workerStats = {};
-        for (const c of doneChunks) {
-            const keys = BigInt('0x' + c.end_hex) - BigInt('0x' + c.start_hex);
-            if (!workerStats[c.worker_name]) workerStats[c.worker_name] = { chunks: 0, keys: 0n };
-            workerStats[c.worker_name].chunks++;
-            workerStats[c.worker_name].keys += keys;
-        }
-
-        const scores = Object.entries(workerStats)
-            .map(([worker_name, { chunks, keys }]) => ({
-                worker_name,
-                completed_chunks: chunks,
-                total_keys: keys.toString()
-            }))
-            .sort((a, b) => {
-                const d = BigInt(b.total_keys) - BigInt(a.total_keys);
-                return d > 0n ? 1 : d < 0n ? -1 : 0;
-            });
-
-        const rawFinders = pid ? db.prepare(`
-            SELECT f.worker_name, f.found_key, f.found_address, f.created_at,
-                c.id AS chunk_global,
-                c.shard_index,
-                c.local_vchunk_start_text AS local_vchunk_start,
-                c.local_vchunk_end_text AS local_vchunk_end,
-                c.vchunk_start,
-                c.vchunk_end
-            FROM findings f
-            JOIN chunks c ON c.id = f.chunk_id
-            WHERE c.puzzle_id = ? AND c.is_test = 0
-            ORDER BY f.id ASC
-        `).all(pid) : [];
-
-        const finders = rawFinders.map(f => {
-            const localStartBig = decodeBigIndex(f.local_vchunk_start, null);
-            const localEndBig = decodeBigIndex(f.local_vchunk_end, null);
-
-            const runStartBig = localStartBig !== null
-                ? localStartBig
-                : decodeBigIndex(f.vchunk_start, null);
-
-            const runEndBig = localEndBig !== null
-                ? localEndBig
-                : decodeBigIndex(f.vchunk_end, null);
-
-            return {
-                worker_name: f.worker_name,
-                found_key: f.found_key,
-                found_address: f.found_address,
-                created_at: f.created_at,
-                chunk_global: f.chunk_global,
-                shard_index: f.shard_index ?? null,
-
-                // keep raw values too, for compatibility if UI still reads them
-                local_vchunk_start: f.local_vchunk_start ?? null,
-                local_vchunk_end: f.local_vchunk_end ?? null,
-                vchunk_start: f.vchunk_start ?? null,
-                vchunk_end: f.vchunk_end ?? null,
-
-                // new normalized display-friendly values
-                vchunk_start_display: runStartBig !== null ? runStartBig.toString() : null,
-                vchunk_end_display: runEndBig !== null ? (runEndBig - 1n).toString() : null,
-                vchunk_run_display: (runStartBig !== null && runEndBig !== null)
-                    ? `${runStartBig.toString()}..${(runEndBig - 1n).toString()}`
-                    : null,
-            };
-        });
-
-        const assignedNow = puzzle ? db.prepare(`
-            SELECT
-                c.id,
-                c.worker_name,
-                c.vchunk_start,
-                c.vchunk_end,
-                c.shard_index,
-                c.local_vchunk_start_text AS local_vchunk_start,
-                c.local_vchunk_end_text AS local_vchunk_end,
-                c.assigned_at,
-                c.heartbeat_at,
-                c.start_hex,
-                c.end_hex
-            FROM chunks c
-            WHERE c.status = 'assigned' AND c.puzzle_id = ? AND c.is_test = 0
-        `).all(puzzle.id) : [];
-
-        const workerAssignedMap = {};
-        for (const c of assignedNow) {
-            const currentJobKeys =
-                c.start_hex && c.end_hex
-                    ? (BigInt('0x' + c.end_hex) - BigInt('0x' + c.start_hex)).toString()
-                    : null;
-
-            const assignedMs = parseUtcDateMs(c.assigned_at);
-            const elapsedSeconds = assignedMs !== null ? Math.max(0, (Date.now() - assignedMs) / 1000) : null;
-
-            const localStartBig = decodeBigIndex(c.local_vchunk_start, null);
-            const localEndBig = decodeBigIndex(c.local_vchunk_end, null);
-
-            const runStartBig = localStartBig !== null
-                ? localStartBig
-                : decodeBigIndex(c.vchunk_start, null);
-
-            const runEndBig = localEndBig !== null
-                ? localEndBig
-                : decodeBigIndex(c.vchunk_end, null);
-
-            workerAssignedMap[c.worker_name] = {
-                current_chunk: c.id,
-                current_shard_index: c.shard_index ?? null,
-                current_vchunk_run: (runStartBig !== null && runEndBig !== null)
-                    ? `${runStartBig.toString()}..${(runEndBig - 1n).toString()}`
-                    : null,
-                current_vchunk_run_start: runStartBig !== null ? runStartBig.toString() : null,
-                current_vchunk_run_end: runEndBig !== null ? runEndBig.toString() : null,
-                assigned_at: c.assigned_at ?? null,
-                heartbeat_at: c.heartbeat_at ?? null,
-                current_job_start_hex: c.start_hex ?? null,
-                current_job_end_hex: c.end_hex ?? null,
-                current_job_keys: currentJobKeys,
-                current_job_elapsed_seconds: elapsedSeconds !== null ? Math.round(elapsedSeconds) : null,
-            };
-        }
-
-        const workers = visibleWorkers.map(w => {
-            const assigned = workerAssignedMap[w.name] || {};
-            return {
-                ...w,
-                fresh: w.fresh === 1,
-                assigned_here: w.assigned_here === 1,
-                active: w.active === 1,
-                current_chunk: assigned.current_chunk ?? null,
-                current_shard_index: assigned.current_shard_index ?? null,
-                current_vchunk_run: assigned.current_vchunk_run ?? null,
-                current_vchunk_run_start: assigned.current_vchunk_run_start ?? null,
-                current_vchunk_run_end: assigned.current_vchunk_run_end ?? null,
-                assigned_at: assigned.assigned_at ?? null,
-                heartbeat_at: assigned.heartbeat_at ?? null,
-                current_job_start_hex: assigned.current_job_start_hex ?? null,
-                current_job_end_hex: assigned.current_job_end_hex ?? null,
-                current_job_keys: assigned.current_job_keys ?? null,
-                current_job_elapsed_seconds: assigned.current_job_elapsed_seconds ?? null,
-                current_job_progress_percent: assigned.assigned_at
-                    ? computeWorkerProgressPercent(assigned.assigned_at, w.hashrate, assigned.current_job_keys)
-                    : null,
-            };
-        });
-
-        let chunks_vis = [];
-        if (puzzle) {
-            const pStart = BigInt('0x' + puzzle.start_hex);
-            const pEnd   = BigInt('0x' + puzzle.end_hex);
-            const pRange = pEnd - pStart;
-
-            const rawChunks = db.prepare(`
-                SELECT id, status, worker_name, start_hex, end_hex, alloc_generation
-                FROM chunks
-                WHERE puzzle_id = ? AND is_test = 0
-                ORDER BY id ASC
-            `).all(puzzle.id);
-
-            chunks_vis = rawChunks.map(c => {
-                const cs = BigInt('0x' + c.start_hex) - pStart;
-                const ce = BigInt('0x' + c.end_hex)   - pStart;
-                return {
-                    id: c.id,
-                    st: c.status,
-                    w:  c.worker_name,
-                    g:  c.alloc_generation || null,
-                    s:  Number(cs * 1000000n / pRange) / 1000000,
-                    e:  Number(ce * 1000000n / pRange) / 1000000,
-                };
-            });
-        }
-
-        const allPuzzles = db.prepare("SELECT id, name, active FROM puzzles ORDER BY id ASC").all();
-
-        const virtualChunksTotal = pid
-            ? (() => {
-                const strategy = puzzle.alloc_strategy || ALLOC_STRATEGY_LEGACY;
-                if (strategy === ALLOC_STRATEGY_VCHUNKS) {
-                    const perShard = decodeBigIndex(
-                        puzzle.virtual_chunks_per_shard_text,
-                        decodeBigIndex(
-                            puzzle.virtual_chunks_per_shard,
-                            decodeBigIndex(
-                                puzzle.virtual_chunk_count_text,
-                                decodeBigIndex(puzzle.virtual_chunk_count, 0n)
-                            )
-                        )
-                    ) || 0n;
-
-                    const shardCount = BigInt(puzzle.shard_count || 0);
-                    return (perShard * shardCount).toString();
-                }
-                return String(db.prepare("SELECT COUNT(*) AS c FROM sectors WHERE puzzle_id = ?").get(pid).c || 0);
-            })()
-            : '0';
-
-        const virtualChunksStarted = pid
-            ? (() => {
-                const strategy = puzzle.alloc_strategy || ALLOC_STRATEGY_LEGACY;
-                if (strategy === ALLOC_STRATEGY_VCHUNKS) {
-                    const rows = db.prepare(`
-                        SELECT local_vchunk_start_text AS local_vchunk_start,
-                        local_vchunk_end_text AS local_vchunk_end,
-                        vchunk_start, vchunk_end
-                        FROM chunks
-                        WHERE puzzle_id = ? AND is_test = 0
-                    `).all(pid);
-                    return sumShardLocalRuns(rows);
-                }
-                return String(db.prepare(`
-                    SELECT COUNT(DISTINCT sector_id) AS c
-                    FROM chunks
-                    WHERE puzzle_id = ? AND is_test = 0 AND sector_id IS NOT NULL
-                `).get(pid).c || 0);
-            })()
-            : '0';
-
-        const virtualChunksCompleted = pid
-            ? (() => {
-                const strategy = puzzle.alloc_strategy || ALLOC_STRATEGY_LEGACY;
-                if (strategy === ALLOC_STRATEGY_VCHUNKS) {
-                    const rows = db.prepare(`
-                        SELECT local_vchunk_start_text AS local_vchunk_start,
-                        local_vchunk_end_text AS local_vchunk_end,
-                        vchunk_start, vchunk_end
-                        FROM chunks
-                        WHERE puzzle_id = ? AND is_test = 0
-                          AND status IN ('completed', 'FOUND')
-                    `).all(pid);
-                    return sumShardLocalRuns(rows);
-                }
-                return String(db.prepare("SELECT COUNT(*) AS c FROM sectors WHERE puzzle_id = ? AND status = 'done'").get(pid).c || 0);
-            })()
-            : '0';
-
-        const allocGenerationRows = pid ? db.prepare(`
-            SELECT alloc_generation, COUNT(*) AS count
-            FROM chunks
-            WHERE puzzle_id = ? AND is_test = 0
-            GROUP BY alloc_generation
-            ORDER BY alloc_generation
-        `).all(pid) : [];
-
-        const allocGenerationCounts = {
-            legacy: 0,
-            affine: 0,
-            feistel: 0,
-        };
-
-        for (const row of allocGenerationRows) {
-            if (row.alloc_generation === 'legacy')  allocGenerationCounts.legacy  = row.count;
-            if (row.alloc_generation === 'affine')  allocGenerationCounts.affine  = row.count;
-            if (row.alloc_generation === 'feistel') allocGenerationCounts.feistel = row.count;
-        }
-
-        res.json({
-            stage: process.env.STAGE || 'PROD',
-            target_minutes: TARGET_MINUTES,
-            timeout_minutes: TIMEOUT_MINUTES,
-            active_minutes: ACTIVE_MINUTES,
-            puzzles: allPuzzles,
-            puzzle: puzzle ? {
-                id: puzzle.id,
-                name: puzzle.name,
-                start_hex: puzzle.start_hex,
-                end_hex: puzzle.end_hex,
-                total_keys: (BigInt('0x' + puzzle.end_hex) - BigInt('0x' + puzzle.start_hex)).toString(),
-                test_chunk: puzzle.test_start_hex ? {
-                    start_hex: puzzle.test_start_hex,
-                    end_hex:   puzzle.test_end_hex,
-                } : null,
-                alloc_strategy: puzzle.alloc_strategy || ALLOC_STRATEGY_LEGACY,
-                alloc_cursor: puzzle.alloc_cursor || 0,
-                virtual_chunk_size_keys: puzzle.virtual_chunk_size_keys || null,
-                virtual_chunk_count: puzzle.virtual_chunk_count_text || puzzle.virtual_chunk_count || null,
-                virtual_chunks_per_shard: puzzle.virtual_chunks_per_shard_text || puzzle.virtual_chunks_per_shard || null,
-                shard_size_keys: puzzle.shard_size_keys || null,
-                shard_count: puzzle.shard_count || null,
-                shard_alloc_cursor: puzzle.shard_alloc_cursor_text || puzzle.shard_alloc_cursor || 0,
-                bootstrap_stage: puzzle.bootstrap_stage || 0,
-            } : null,
-            active_workers_count: visibleWorkers.filter(w => w.active).length,
-            inactive_workers_count: visibleWorkers.filter(w => !w.active).length,
-            total_hashrate: totalHashrate,
-            completed_chunks: completedChunks,
-            reclaimed_chunks: reclaimedChunks,
-            total_keys_completed: totalKeysCompleted.toString(),
-            virtual_chunks: {
-                total: virtualChunksTotal,
-                started: virtualChunksStarted,
-                completed: virtualChunksCompleted,
-            },
-            // Backward-compatible alias for the old dashboard
-            shards: {
-                total: virtualChunksTotal,
-                started: virtualChunksStarted,
-                completed: virtualChunksCompleted,
-            },
-            workers,
-            scores,
-            finders,
-            chunks_vis,
-            alloc_generations: allocGenerationCounts,
-        });
-    });
-
-    app.post('/api/v1/heartbeat', (req, res) => {
-        const { name, job_id } = req.body;
-        if (!name || job_id === null || job_id === undefined) {
-            return res.status(400).json({ error: 'Missing name or job_id' });
-        }
-
-        db.prepare(`
-            INSERT INTO workers (name, hashrate, last_seen)
-            VALUES (?, 0, CURRENT_TIMESTAMP)
-            ON CONFLICT(name) DO UPDATE SET last_seen = CURRENT_TIMESTAMP
-        `).run(name);
-
-        db.prepare(`
-            UPDATE chunks
-            SET heartbeat_at = CURRENT_TIMESTAMP
-            WHERE id = ? AND worker_name = ? AND status = 'assigned'
-        `).run(job_id, name);
-
-        res.json({ ok: true });
-    });
-
-    app.post('/api/v1/admin/set-puzzle', (req, res) => {
-        const {
-            name,
-            start_hex,
-            end_hex,
-            alloc_strategy,
-            alloc_seed,
-            virtual_chunk_size_keys
-        } = req.body;
-
-        if (!name || !start_hex || !end_hex) {
-            return res.status(400).json({ error: 'Missing name, start_hex, or end_hex' });
-        }
-        if (!isValidHex(start_hex) || !isValidHex(end_hex)) {
-            return res.status(400).json({ error: 'start_hex and end_hex must be valid hex strings' });
-        }
-
-        const strategy = alloc_strategy || DEFAULT_ALLOC_STRATEGY;
-        if (strategy !== ALLOC_STRATEGY_LEGACY && strategy !== ALLOC_STRATEGY_VCHUNKS) {
-            return res.status(400).json({ error: `alloc_strategy must be ${ALLOC_STRATEGY_LEGACY} or ${ALLOC_STRATEGY_VCHUNKS}` });
-        }
-
-        const startNorm = start_hex.replace(/^0x/i, '').padStart(64, '0').toLowerCase();
-        const endNorm   = end_hex.replace(/^0x/i, '').padStart(64, '0').toLowerCase();
-
-        if (BigInt('0x' + endNorm) <= BigInt('0x' + startNorm)) {
-            return res.status(400).json({ error: 'end_hex must be greater than start_hex' });
-        }
-
-        const puzzleRange = BigInt('0x' + endNorm) - BigInt('0x' + startNorm);
-
-        let vchunkSize = null;
-        if (strategy === ALLOC_STRATEGY_VCHUNKS) {
-            const shardSize = bigIntMin(chooseShardSize(puzzleRange), puzzleRange);
-            const parsed = parsePositiveBigInt(virtual_chunk_size_keys, null);
-            vchunkSize = parsed ? bigIntMin(parsed, shardSize) : chooseVirtualChunkSizeForShard(shardSize);
-            validateVirtualChunkCountOrThrow(ceilDiv(shardSize, vchunkSize), 'set-puzzle');
-        }
-
-        db.transaction(() => {
-            db.prepare("UPDATE puzzles SET active = 0").run();
-
-            const existing = db.prepare("SELECT * FROM puzzles WHERE name = ?").get(name);
-            let puzzleId;
-
-            const seed = alloc_seed || defaultAllocSeedForPuzzle({ name, start_hex: startNorm, end_hex: endNorm }, strategy);
-            const sameSeed = !existing || existing.alloc_seed === seed;
-            const sameVChunkSize = !existing || !existing.virtual_chunk_size_keys || existing.virtual_chunk_size_keys === (vchunkSize ? vchunkSize.toString() : null);
-
-            if (
-                existing &&
-                existing.start_hex === startNorm &&
-                existing.end_hex === endNorm &&
-                (existing.alloc_strategy || ALLOC_STRATEGY_LEGACY) === strategy &&
-                sameSeed &&
-                sameVChunkSize
-            ) {
-                db.prepare("UPDATE puzzles SET active = 1 WHERE id = ?").run(existing.id);
-                puzzleId = existing.id;
-                ensureAllocatorForPuzzle(db, puzzleId);
-            } else {
-                const info = db.prepare(`
-                    INSERT INTO puzzles (
-                        name, start_hex, end_hex, active,
-                        alloc_strategy, alloc_seed, alloc_cursor,
-                        virtual_chunk_size_keys, virtual_chunk_count, bootstrap_stage
-                    )
-                    VALUES (?, ?, ?, 1, ?, ?, 0, ?, NULL, 0)
-                `).run(
-                    name,
-                    startNorm,
-                    endNorm,
-                    strategy,
-                    seed,
-                    vchunkSize ? vchunkSize.toString() : null
-                );
-
-                puzzleId = info.lastInsertRowid;
-
-                if (strategy === ALLOC_STRATEGY_VCHUNKS) {
-                    seedVirtualChunks(db, puzzleId, startNorm, endNorm, seed, vchunkSize);
-                } else {
-                    seedSectors(db, puzzleId, startNorm, endNorm);
-                }
-            }
-        })();
-
-        const puzzle = db.prepare("SELECT * FROM puzzles WHERE active = 1 LIMIT 1").get();
-        console.log(`[Admin] Active puzzle set to: ${puzzle.name} [${puzzle.start_hex} .. ${puzzle.end_hex}] strategy=${puzzle.alloc_strategy || ALLOC_STRATEGY_LEGACY}`);
-        res.json({ ok: true, puzzle });
-    });
-
-    app.post('/api/v1/admin/set-test-chunk', (req, res) => {
-        const puzzle = db.prepare("SELECT * FROM puzzles WHERE active = 1 LIMIT 1").get();
-        if (!puzzle) return res.status(503).json({ error: 'No active puzzle' });
-
-        const { start_hex, end_hex } = req.body;
-
-        if (!start_hex) {
-            db.prepare("UPDATE puzzles SET test_start_hex = NULL, test_end_hex = NULL WHERE id = ?").run(puzzle.id);
-            console.log(`[Admin] Test chunk cleared for puzzle ${puzzle.name}`);
-            return res.json({ ok: true, test_chunk: null });
-        }
-
-        if (!isValidHex(start_hex)) {
-            return res.status(400).json({ error: 'start_hex must be a valid hex string' });
-        }
-
-        const startNorm = start_hex.replace(/^0x/i, '').padStart(64, '0').toLowerCase();
-        const ts = BigInt('0x' + startNorm);
-
-        let endNorm;
-        if (end_hex) {
-            if (!isValidHex(end_hex)) {
-                return res.status(400).json({ error: 'end_hex must be a valid hex string' });
-            }
-            endNorm = end_hex.replace(/^0x/i, '').padStart(64, '0').toLowerCase();
-        } else {
-            const MAX_256 = (1n << 256n) - 1n;
-            if (ts + GPU_BATCH_KEYS > MAX_256) {
-                return res.status(400).json({ error: 'auto-resolved end_hex exceeds 256-bit range' });
-            }
-            endNorm = (ts + GPU_BATCH_KEYS).toString(16).padStart(64, '0');
-        }
-
-        const te = BigInt('0x' + endNorm);
-        if (te <= ts) {
-            return res.status(400).json({ error: 'end_hex must be greater than start_hex' });
-        }
-
-        db.prepare(`
-            UPDATE puzzles SET test_start_hex = ?, test_end_hex = ?
-            WHERE id = ?
-        `).run(startNorm, endNorm, puzzle.id);
-
-        console.log(`[Admin] Test chunk set: ${startNorm} .. ${endNorm}`);
-        res.json({ ok: true, test_chunk: { start_hex: startNorm, end_hex: endNorm } });
-    });
-
-    app.post('/api/v1/admin/activate-puzzle', (req, res) => {
-        const { id } = req.body;
-        if (!id) return res.status(400).json({ error: 'Missing id' });
-
-        const target = db.prepare("SELECT * FROM puzzles WHERE id = ?").get(id);
-        if (!target) return res.status(404).json({ error: 'Puzzle not found' });
-
-        db.transaction(() => {
-            db.prepare("UPDATE puzzles SET active = 0").run();
-            db.prepare("UPDATE puzzles SET active = 1 WHERE id = ?").run(id);
-        })();
-
-        ensureAllocatorForPuzzle(db, target.id);
-
-        const puzzle = db.prepare("SELECT * FROM puzzles WHERE id = ?").get(id);
-        console.log(`[Admin] Active puzzle switched to: ${puzzle.name}`);
-        res.json({ ok: true, puzzle });
-    });
-
-    app.get('/api/v1/admin/puzzles', (req, res) => {
-        const puzzles = db.prepare(`
-            SELECT id, name, active, start_hex, end_hex,
-                   alloc_strategy, alloc_seed, alloc_cursor,
-                   shard_size_keys, shard_count, shard_alloc_cursor, shard_alloc_cursor_text,
-                   virtual_chunk_size_keys, virtual_chunk_count, virtual_chunk_count_text,
-                   virtual_chunks_per_shard, virtual_chunks_per_shard_text,
-                   bootstrap_stage
-            FROM puzzles
-            ORDER BY id ASC
-        `).all();
-        res.json({ puzzles });
-    });
-
-    return app;
-}
-
-// --- Production entry point ---
-if (require.main === module) {
-    const db = new Database(DB_PATH);
-    db.pragma('journal_mode = WAL');
-
-    db.exec(`
-      CREATE TABLE IF NOT EXISTS puzzles (
-        id INTEGER PRIMARY KEY AUTOINCREMENT,
-        name TEXT NOT NULL,
-        start_hex TEXT NOT NULL,
-        end_hex TEXT NOT NULL,
-        active INTEGER NOT NULL DEFAULT 0
-      );
-
-      CREATE TABLE IF NOT EXISTS workers (
-        name TEXT PRIMARY KEY,
-        hashrate REAL,
-        last_seen DATETIME DEFAULT CURRENT_TIMESTAMP
-      );
-
-      CREATE TABLE IF NOT EXISTS chunks (
-        id INTEGER PRIMARY KEY AUTOINCREMENT,
-        puzzle_id INTEGER,
-        start_hex TEXT,
-        end_hex TEXT,
-        status TEXT,
-        worker_name TEXT,
-        prev_worker_name TEXT,
-        assigned_at DATETIME,
-        heartbeat_at DATETIME,
-        found_key TEXT,
-        found_address TEXT,
-        is_test INTEGER NOT NULL DEFAULT 0,
-        sector_id INTEGER
-      );
-
-      CREATE INDEX IF NOT EXISTS idx_chunks_puzzle_status ON chunks (puzzle_id, status);
-
-      CREATE TABLE IF NOT EXISTS sectors (
-        id          INTEGER PRIMARY KEY AUTOINCREMENT,
-        puzzle_id   INTEGER NOT NULL,
-        start_hex   TEXT NOT NULL,
-        end_hex     TEXT NOT NULL,
-        current_hex TEXT NOT NULL,
-        status      TEXT NOT NULL DEFAULT 'open'
-      );
-
-      CREATE INDEX IF NOT EXISTS idx_sectors_puzzle_status ON sectors (puzzle_id, status);
-      CREATE INDEX IF NOT EXISTS idx_sectors_puzzle_id     ON sectors (puzzle_id, id);
-      CREATE UNIQUE INDEX IF NOT EXISTS idx_sectors_unique_span ON sectors (puzzle_id, start_hex, end_hex);
-
-      CREATE TABLE IF NOT EXISTS findings (
-        id INTEGER PRIMARY KEY AUTOINCREMENT,
-        chunk_id INTEGER NOT NULL,
-        worker_name TEXT NOT NULL,
-        found_key TEXT NOT NULL,
-        found_address TEXT,
-        created_at DATETIME DEFAULT CURRENT_TIMESTAMP
-      );
-
-      CREATE UNIQUE INDEX IF NOT EXISTS idx_findings_dedup ON findings (chunk_id, worker_name, found_key);
-
-        CREATE TABLE IF NOT EXISTS puzzle_shards (
-            puzzle_id INTEGER NOT NULL,
-            shard_index INTEGER NOT NULL,
-            status TEXT NOT NULL DEFAULT 'open',
-            alloc_cursor_text TEXT NOT NULL DEFAULT '0000000000000000000000000000000000000000000000000000000000000000',
-            bootstrap_stage INTEGER NOT NULL DEFAULT 0,
-            assigned_runs INTEGER NOT NULL DEFAULT 0,
-            completed_runs INTEGER NOT NULL DEFAULT 0,
-            PRIMARY KEY (puzzle_id, shard_index)
-        );
-    `);
-
-    try { db.prepare("ALTER TABLE puzzles ADD COLUMN test_start_hex TEXT").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE puzzles ADD COLUMN test_end_hex TEXT").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE puzzles ADD COLUMN alloc_strategy TEXT").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE puzzles ADD COLUMN alloc_seed TEXT").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE puzzles ADD COLUMN alloc_cursor INTEGER NOT NULL DEFAULT 0").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE puzzles ADD COLUMN virtual_chunk_size_keys TEXT").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE puzzles ADD COLUMN virtual_chunk_count INTEGER").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE puzzles ADD COLUMN bootstrap_stage INTEGER NOT NULL DEFAULT 0").run(); } catch (_) {}
-
-    try { db.prepare("ALTER TABLE puzzles ADD COLUMN shard_size_keys TEXT").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE puzzles ADD COLUMN shard_count INTEGER").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE puzzles ADD COLUMN shard_alloc_cursor INTEGER NOT NULL DEFAULT 0").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE puzzles ADD COLUMN virtual_chunks_per_shard INTEGER").run(); } catch (_) {}
-
-    try { db.prepare("ALTER TABLE workers ADD COLUMN version TEXT").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE workers ADD COLUMN min_chunk_keys TEXT").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE workers ADD COLUMN chunk_quantum_keys TEXT").run(); } catch (_) {}
-
-    try { db.prepare("ALTER TABLE chunks ADD COLUMN is_test INTEGER NOT NULL DEFAULT 0").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE chunks ADD COLUMN prev_worker_name TEXT").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE chunks ADD COLUMN alloc_block_id INTEGER").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE chunks ADD COLUMN vchunk_start INTEGER").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE chunks ADD COLUMN vchunk_end INTEGER").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE chunks ADD COLUMN heartbeat_at DATETIME").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE chunks ADD COLUMN alloc_generation TEXT").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE chunks ADD COLUMN shard_index INTEGER").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE chunks ADD COLUMN local_vchunk_start INTEGER").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE chunks ADD COLUMN local_vchunk_end INTEGER").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE puzzles ADD COLUMN shard_alloc_cursor_text TEXT").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE puzzles ADD COLUMN virtual_chunk_count_text TEXT").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE puzzles ADD COLUMN virtual_chunks_per_shard_text TEXT").run(); } catch (_) {}
-
-    try { db.prepare("ALTER TABLE chunks ADD COLUMN local_vchunk_start_text TEXT").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE chunks ADD COLUMN local_vchunk_end_text TEXT").run(); } catch (_) {}
-    try { db.prepare("ALTER TABLE puzzle_shards ADD COLUMN alloc_cursor_text TEXT").run(); } catch (_) {}
-    try {
-        db.prepare("UPDATE chunks SET heartbeat_at = assigned_at WHERE heartbeat_at IS NULL AND assigned_at IS NOT NULL").run();
-    } catch (_) {}
-
-    try {
-    db.prepare(`
-        UPDATE puzzles
-        SET shard_alloc_cursor_text = CASE
-            WHEN shard_alloc_cursor_text IS NOT NULL AND shard_alloc_cursor_text != '' THEN shard_alloc_cursor_text
-            WHEN shard_alloc_cursor IS NOT NULL THEN printf('%064x', shard_alloc_cursor)
-            ELSE '0000000000000000000000000000000000000000000000000000000000000000'
-        END
-        WHERE shard_alloc_cursor_text IS NULL OR shard_alloc_cursor_text = ''
-    `).run();
-    } catch (_) {}
-
-    try {
-        db.prepare(`
-            UPDATE puzzles
-            SET virtual_chunk_count_text = CASE
-                WHEN virtual_chunk_count_text IS NOT NULL AND virtual_chunk_count_text != '' THEN virtual_chunk_count_text
-                WHEN virtual_chunk_count IS NOT NULL THEN CAST(virtual_chunk_count AS TEXT)
-                ELSE NULL
-            END
-            WHERE virtual_chunk_count_text IS NULL OR virtual_chunk_count_text = ''
-        `).run();
-    } catch (_) {}
-
-    try {
-        db.prepare(`
-            UPDATE puzzles
-            SET virtual_chunks_per_shard_text = CASE
-                WHEN virtual_chunks_per_shard_text IS NOT NULL AND virtual_chunks_per_shard_text != '' THEN virtual_chunks_per_shard_text
-                WHEN virtual_chunks_per_shard IS NOT NULL THEN CAST(virtual_chunks_per_shard AS TEXT)
-                ELSE NULL
-            END
-            WHERE virtual_chunks_per_shard_text IS NULL OR virtual_chunks_per_shard_text = ''
-        `).run();
-    } catch (_) {}
-
-    try {
-        db.prepare(`
-            UPDATE chunks
-            SET local_vchunk_start_text = CASE
-                WHEN local_vchunk_start_text IS NOT NULL AND local_vchunk_start_text != '' THEN local_vchunk_start_text
-                WHEN local_vchunk_start IS NOT NULL THEN printf('%064x', local_vchunk_start)
-                ELSE NULL
-            END
-            WHERE local_vchunk_start_text IS NULL OR local_vchunk_start_text = ''
-        `).run();
-    } catch (_) {}
-
-    try {
-        db.prepare(`
-            UPDATE chunks
-            SET local_vchunk_end_text = CASE
-                WHEN local_vchunk_end_text IS NOT NULL AND local_vchunk_end_text != '' THEN local_vchunk_end_text
-                WHEN local_vchunk_end IS NOT NULL THEN printf('%064x', local_vchunk_end)
-                ELSE NULL
-            END
-            WHERE local_vchunk_end_text IS NULL OR local_vchunk_end_text = ''
-        `).run();
-    } catch (_) {}
-
-    try {
-        const shardCols = db.prepare(`PRAGMA table_info(puzzle_shards)`).all();
-        const shardColNames = new Set(shardCols.map(c => c.name));
-
-        if (shardColNames.has('alloc_cursor')) {
-            db.prepare(`
-                UPDATE puzzle_shards
-                SET alloc_cursor_text = CASE
-                    WHEN alloc_cursor_text IS NOT NULL AND alloc_cursor_text != '' THEN alloc_cursor_text
-                    WHEN alloc_cursor IS NOT NULL THEN printf('%064x', alloc_cursor)
-                    ELSE '0000000000000000000000000000000000000000000000000000000000000000'
-                END
-                WHERE alloc_cursor_text IS NULL OR alloc_cursor_text = ''
-            `).run();
-        } else {
-            db.prepare(`
-                UPDATE puzzle_shards
-                SET alloc_cursor_text = '0000000000000000000000000000000000000000000000000000000000000000'
-                WHERE alloc_cursor_text IS NULL OR alloc_cursor_text = ''
-            `).run();
-        }
-    } catch (_) {}
-
-    try { db.prepare("CREATE UNIQUE INDEX IF NOT EXISTS idx_findings_dedup ON findings (chunk_id, worker_name, found_key)").run(); } catch (_) {}
-    try { db.prepare("CREATE INDEX IF NOT EXISTS idx_chunks_vchunk_span ON chunks (puzzle_id, vchunk_start, vchunk_end, status)").run(); } catch (_) {}
-
-    try { db.prepare("CREATE INDEX IF NOT EXISTS idx_puzzle_shards_status ON puzzle_shards (puzzle_id, status, shard_index)").run(); } catch (_) {}
-    try { db.prepare("CREATE INDEX IF NOT EXISTS idx_chunks_shard_span ON chunks (puzzle_id, shard_index, local_vchunk_start, local_vchunk_end, status)").run(); } catch (_) {}
-
-    try { db.prepare("CREATE INDEX IF NOT EXISTS idx_chunks_shard_span_text ON chunks (puzzle_id, shard_index, local_vchunk_start_text, local_vchunk_end_text, status)").run(); } catch (_) {}
-
-    try {
-        db.prepare(`
-            UPDATE chunks SET is_test = 1
-            WHERE is_test = 0
-              AND EXISTS (
-                SELECT 1 FROM puzzles p
-                WHERE p.id = chunks.puzzle_id
-                  AND p.test_start_hex IS NOT NULL
-                  AND chunks.start_hex = p.test_start_hex
-                  AND chunks.end_hex   = p.test_end_hex
-              )
-        `).run();
-    } catch (_) {}
-
-    // Normalize allocator defaults on pre-existing puzzles
-    try {
-        db.prepare(`
-            UPDATE puzzles
-            SET alloc_strategy = ?
-            WHERE alloc_strategy IS NULL
-        `).run(ALLOC_STRATEGY_LEGACY);
-    } catch (_) {}
-
-    try {
-        db.prepare(`
-            UPDATE puzzle_shards
-            SET alloc_cursor_text = '0000000000000000000000000000000000000000000000000000000000000000'
-            WHERE alloc_cursor_text IS NULL OR alloc_cursor_text = ''
-        `).run();
-    } catch (_) {}
-
-    // Seed from KEYSPACE_<NAME>=<start_hex>:<end_hex>
-    for (const [key, value] of Object.entries(process.env)) {
-        if (!key.startsWith('KEYSPACE_')) continue;
-
-        const name = key.slice('KEYSPACE_'.length).replace(/_/g, ' ');
-        const [startRaw, endRaw] = (value || '').split(':');
-
-        if (!startRaw || !endRaw || !isValidHex(startRaw) || !isValidHex(endRaw)) {
-            console.warn(`[Config] Skipping invalid keyspace ${key}=${value} — expected start_hex:end_hex`);
-            continue;
-        }
-
-        const startNorm = startRaw.replace(/^0x/i, '').padStart(64, '0').toLowerCase();
-        const endNorm   = endRaw.replace(/^0x/i, '').padStart(64, '0').toLowerCase();
-
-        if (BigInt('0x' + endNorm) <= BigInt('0x' + startNorm)) {
-            console.warn(`[Config] Skipping ${key} — end_hex must be greater than start_hex`);
-            continue;
-        }
-
-        const existing = db.prepare("SELECT id FROM puzzles WHERE name = ?").get(name);
-        if (!existing) {
-            const strategy = DEFAULT_ALLOC_STRATEGY;
-            const seed = defaultAllocSeedForPuzzle({ name, start_hex: startNorm, end_hex: endNorm }, strategy);
-
-            let virtualChunkSize = null;
-            if (strategy === ALLOC_STRATEGY_VCHUNKS) {
-                const range = BigInt('0x' + endNorm) - BigInt('0x' + startNorm);
-                const shardSize = bigIntMin(chooseShardSize(range), range);
-                virtualChunkSize = chooseVirtualChunkSizeForShard(shardSize);
-            }
-
-            const info = db.prepare(`
-                INSERT INTO puzzles (
-                    name, start_hex, end_hex, active,
-                    alloc_strategy, alloc_seed, alloc_cursor,
-                    virtual_chunk_size_keys, virtual_chunk_count, bootstrap_stage
-                )
-                VALUES (?, ?, ?, 0, ?, ?, 0, ?, NULL, 0)
-            `).run(
-                name,
-                startNorm,
-                endNorm,
-                strategy,
-                seed,
-                virtualChunkSize ? virtualChunkSize.toString() : null
-            );
-
-            console.log(`[Config] Seeded keyspace: ${name}`);
-            if (strategy === ALLOC_STRATEGY_VCHUNKS) {
-                seedVirtualChunks(db, info.lastInsertRowid, startNorm, endNorm, seed, virtualChunkSize);
-            } else {
-                seedSectors(db, info.lastInsertRowid, startNorm, endNorm);
-            }
-        }
-    }
-
-    const puzzleCount = db.prepare("SELECT COUNT(*) AS count FROM puzzles").get().count;
-    if (puzzleCount === 0) {
-        const startHex = '0400000000000000000'.padStart(64, '0');
-        const endHex   = '07fffffffffffffffff'.padStart(64, '0');
-        const name = 'Puzzle #71';
-        const strategy = DEFAULT_ALLOC_STRATEGY;
-        const seed = defaultAllocSeedForPuzzle({ name, start_hex: startHex, end_hex: endHex }, strategy);
-
-        let virtualChunkSize = null;
-        if (strategy === ALLOC_STRATEGY_VCHUNKS) {
-            const range = BigInt('0x' + endHex) - BigInt('0x' + startHex);
-            const shardSize = bigIntMin(chooseShardSize(range), range);
-            virtualChunkSize = chooseVirtualChunkSizeForShard(shardSize);
-        }
-
-        const info = db.prepare(`
-            INSERT INTO puzzles (
-                name, start_hex, end_hex, active,
-                alloc_strategy, alloc_seed, alloc_cursor,
-                virtual_chunk_size_keys, virtual_chunk_count, bootstrap_stage
-            )
-            VALUES (?, ?, ?, 1, ?, ?, 0, ?, NULL, 0)
-        `).run(
-            name,
-            startHex,
-            endHex,
-            strategy,
-            seed,
-            virtualChunkSize ? virtualChunkSize.toString() : null
-        );
-
-        console.log('[Init] Seeded Puzzle #71 as active puzzle.');
-        if (strategy === ALLOC_STRATEGY_VCHUNKS) {
-            seedVirtualChunks(db, info.lastInsertRowid, startHex, endHex, seed, virtualChunkSize);
-        } else {
-            seedSectors(db, info.lastInsertRowid, startHex, endHex);
-        }
-    }
-
-    const activeCount = db.prepare("SELECT COUNT(*) AS count FROM puzzles WHERE active = 1").get().count;
-    if (activeCount === 0) {
-        db.prepare("UPDATE puzzles SET active = 1 WHERE id = (SELECT MIN(id) FROM puzzles)").run();
-        console.log('[Init] No active puzzle found — activated the first one.');
-    } else if (activeCount > 1) {
-        db.prepare("UPDATE puzzles SET active = 0 WHERE id != (SELECT MAX(id) FROM puzzles WHERE active = 1)").run();
-        console.log('[Init] Multiple active puzzles found — deactivated all but the latest.');
-    }
-
-    const allPuzzles = db.prepare("SELECT id FROM puzzles").all();
-    for (const p of allPuzzles) {
-        try {
-            ensureAllocatorForPuzzle(db, p.id);
-        } catch (e) {
-            console.error(`[Init] Failed to ensure allocator for puzzle ${p.id}: ${e.message}`);
-            throw e;
-        }
-    }
-
-    const app = createApp(db);
-
-    setInterval(() => {
-        const info = db.prepare(`
-            UPDATE chunks
-            SET status = 'reclaimed',
-                prev_worker_name = worker_name,
-                worker_name = NULL,
-                assigned_at = NULL,
-                heartbeat_at = NULL
-            WHERE status = 'assigned'
-              AND COALESCE(heartbeat_at, assigned_at) < datetime('now', '-${TIMEOUT_MINUTES} minutes')
-        `).run();
-
-        if (info.changes > 0) {
-            console.log(`[System] Reclaimed ${info.changes} abandoned chunks back to the pool.`);
-        }
-    }, 60000);
-
-    app.listen(PORT, '127.0.0.1', () => {
-        console.log(`[puzzpool] server running on http://127.0.0.1:${PORT}`);
-        console.log(`[puzzpool] database: ${DB_PATH}`);
-        if (process.env.ADMIN_TOKEN) console.log('[puzzpool] admin token auth: enabled');
-    });
-}
-
-module.exports = {
-    createApp,
-    isValidHex,
-    randomBigIntInRange,
-    normalizeHashrate,
-    seedSectors,
-    seedVirtualChunks,
-    deriveAffinePermutationParams,
-    permuteIndexAffine,
-    defaultAllocSeedForPuzzle,
-    bitLengthBigInt,
-    nextEven,
-    feistelRoundValue,
-    permutePow2Feistel,
-    permuteIndexFeistel,
-    PERMUTATION_MODE,
-    ACTIVE_MINUTES,
-    REACTIVATE_MINUTES,
-    ALLOC_STRATEGY_LEGACY,
-    ALLOC_STRATEGY_VCHUNKS,
-    DEFAULT_VIRTUAL_CHUNK_SIZE_KEYS,
-    MAX_STORED_VCHUNKS,
-    MAX_ALLOC_PROBES,
-    nextPowerOfTwoBigInt,
-    chooseShardSize,
-    chooseVirtualChunkSizeForShard,
-    shardLocalVirtualChunkRangeToHex,
-    TARGET_MAX_SHARDS,
-    TARGET_MAX_VCHUNKS_PER_SHARD,
-    PUZZLE71_RANGE_FLOOR,
-};
